% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{centroids}
\alias{centroids}
\alias{centroids,HiCDOCDataSet-method}
\title{Extracts centroids}
\usage{
\S4method{centroids}{HiCDOCDataSet}(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
A \code{tibble} object of the centroids
}
\description{
This function extracts the centroids from \code{\link{HiCDOCDataSet}}.
}
\examples{
exp <- HiCDOCExample()
exp <- HiCDOC(exp)
centroids(exp)

}
