% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkParameters}
\alias{checkParameters}
\title{Check object@parameters, return default if NULL}
\usage{
checkParameters(parameters, paramnames)
}
\arguments{
\item{parameters}{list of parameters}

\item{paramnames}{character vector, names of parameters to check}
}
\value{
list of updated parameters, default from HiCDOCDefaultParameters()
if null.
}
\description{
Check object@parameters, return default if NULL
}
