% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{chromosomes}
\alias{chromosomes}
\alias{chromosomes,HiCDOCDataSet-method}
\title{Accessors for the 'chromosomes' slot of an HiCDOCDataSet object}
\usage{
\S4method{chromosomes}{HiCDOCDataSet}(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
A character vector
}
\description{
The \code{chromosomes} slot contains the names of the chromosomes.
}
\examples{
exp <- HiCDOCExample()
chromosomes(exp)

}
