% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{clusterize}
\alias{clusterize}
\title{Segregate genomic positions into 2 clusters using constrained k-means.}
\usage{
clusterize(object, parallel = FALSE, nbCores = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{parallel}{Logical. Should parallel processing be used?}

\item{nbCores}{Integer, the number of cores to use for parallelization.}
}
\value{
A \code{HiCDOCDataSet} object, with:
- centroid (vector) for each cluster
- compartment (cluster number) for each genomic position
- distances to centroids (float) for each genomic position in each replicate
- concordance (float in [-1, 1]) for each genomic position in each replicate
}
\description{
Segregate genomic positions into 2 clusters using constrained k-means.
}
