% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{clusterizeChrCond}
\alias{clusterizeChrCond}
\title{Segregate genomic positions into 2 clusters using constrained k-means.}
\usage{
clusterizeChrCond(object, chromosomeId, conditionId)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{chromosomeId}{A character or numeric value.
Name or number of the chromosome}

\item{conditionId}{A character or numeric value.
Name or number of the condition}
}
\value{
A \code{HiCDOCDataSet} object, with:
- centroid (vector) for each cluster
- compartment (cluster number) for each genomic position
- distances to centroids (float) for each genomic position in each replicate
- concordance (float in [-1, 1]) for each genomic position in each replicate
}
\description{
Segregate genomic positions into 2 clusters using constrained k-means.
}
