% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{compartments}
\alias{compartments}
\alias{compartments,HiCDOCDataSet-method}
\title{Extracts compartments}
\usage{
\S4method{compartments}{HiCDOCDataSet}(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
A \code{tibble} object of the compartments
}
\description{
This function extracts the compartments from \code{\link{HiCDOCDataSet}}.
}
\examples{
exp <- HiCDOCExample()
exp <- HiCDOC(exp)
compartments(exp)
}
