% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpersConstructor.R
\name{computeBinSize}
\alias{computeBinSize}
\title{Compute the binSize of an HiCDOCDataset object}
\usage{
computeBinSize(object)
}
\arguments{
\item{object}{a HiCDOCDataSet object.}
}
\value{
binSize, an integer.
}
\description{
The function takes a HiCDOCDataset object and compute the binSize from
the positions in the interactions matrix.
It takes the minimum of difference between position.1 and position.2
}
