% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{concordances}
\alias{concordances}
\alias{concordances,HiCDOCExp-method}
\title{Extracts concordances}
\usage{
\S4method{concordances}{HiCDOCExp}(object)
}
\arguments{
\item{object}{An \code{HiCDOCExp} object.}
}
\value{
A \code{tibble} object of the concordance
}
\description{
This function extracts the concordances from \code{\link{HiCDOCExp}}.
}
\examples{
exp <- HiCDOCExample()
exp <- HiCDOC(exp)
concordances(exp)

}
