% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpersConstructor.R
\name{determinePositions}
\alias{determinePositions}
\title{Determine the positions corresponding to the bins of a HiCDOCDataSet object}
\usage{
determinePositions(object)
}
\arguments{
\item{object}{a HiCDOCDataSet}
}
\value{
a tibble, with 4 columns : chromosome, bin, start, end.
}
\description{
The function determines, for each chromosome, the positions corresponding
to the bins from the interaction matrix and the binSize. The positions will
be stored in positions(object).
}
