% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleHiCDOCDataSet}
\alias{exampleHiCDOCDataSet}
\title{Example HiCDOCDataSet.}
\format{
S4 HiCDOCDataSet object with the following characteristics:
\describe{
  \item{chromosomes}{4 chromosomes: W, X, Y, Z}
  \item{conditions}{3 conditions: 1, 2, 3}
  \item{replicates}{3 replicates: R1, R2, R3}
  \item{binSize}{A resolution of 137 bases}
}
}
\usage{
data(exampleHiCDOCDataSet)
}
\value{
A \code{\link{HiCDOCDataSet}}.
}
\description{
A S4 HiCDOCDataSet object with 4 chromosomes, 3 conditions and 3 replicates.
}
\examples{
data(exampleHiCDOCDataSet)
exampleHiCDOCDataSet
}
\keyword{datasets}
