% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompartmentChanges.R
\name{extract_legends}
\alias{extract_legends}
\title{Function to extract legends from ggplot2 objects}
\usage{
extract_legends(x)
}
\arguments{
\item{x}{a grob}
}
\value{
the legend as a grob
}
\description{
Inspired from gtable::gtable_filter (not used to avoid dependency)
}
\keyword{internal}
