% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterWeakPositions.R
\name{filterWeakChr}
\alias{filterWeakChr}
\title{Look for the weak positions for a given chromosome}
\usage{
filterWeakChr(
  interChr,
  totalBinsChr,
  binsize,
  replicates,
  conditions,
  threshold = 0
)
}
\arguments{
\item{binsize}{Size of the bins in the HiCDOC object}

\item{replicates}{Levels of replicates. Given from object@conditions}

\item{conditions}{Levels of conditions (repeated by replicates).
Given from object@conditions}

\item{threshold}{Threshold for the selection of "empty". The function will
return the bin if the mean of value if inferior to `threshold`,
for at least one condition and a replicate (after we reconstruct the 0 values).
Default to 0.}

\item{inter_chr}{Interaction matrix for the chromosome}

\item{totalBins_chr}{Number of total bins for this chromosome}
}
\value{

}
\description{
The function identifies the weak positions from the interaction matrix
for a given chromosome., in a recursive way.
}
\examples{
object <- HiCDOCExample()
chr <- object@chromosomes[[1]]
filterWeakChr(object@interactions[object@interactions$chromosome == chr,],
             object@totalBins[[chr]],
             object@binSize,
             object@replicates,
             object@conditions,
             object@filterThreshold)
}
