% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{findMissingCondLevels}
\alias{findMissingCondLevels}
\title{Find unbalanced condition levels}
\usage{
findMissingCondLevels(conditions, replicates)
}
\arguments{
\item{conditions}{conditions vector, with repetitions, as in object@conditions}

\item{replicates}{replicates vector, with repetitions, as in object@conditions}
}
\value{
Numerical vector of the same length as conditions, giving the number of 
missing repetitions to reach a balance between conditions.
}
\description{
Find unbalanced condition levels
}
