% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fullInteractions}
\alias{fullInteractions}
\title{Full Interactions Matrix}
\usage{
fullInteractions(object)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object}
}
\value{
a tibble, long version of the full interactions matrix.
}
\description{
From an interactions matrix, that can be sparse and an upper of lower matrix, 
generate the full interactions matrix, symetric and filling gaps with 0.
}
