% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{h5readCatch}
\alias{h5readCatch}
\title{Read an HDF5 file, and extract a field.
It raises an exception if the field is not present.}
\usage{
h5readCatch(file, name)
}
\arguments{
\item{file}{The name of a file in HDF5 format.}

\item{name}{The name a field name.}
}
\value{
The value associated with the field.
}
\description{
Read an HDF5 file, and extract a field.
It raises an exception if the field is not present.
}
