% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{HiCDOCDataSet-methods}
\alias{HiCDOCDataSet-methods}
\alias{chromosomes}
\alias{conditions}
\alias{replicates}
\alias{resolution}
\alias{interactions}
\alias{compartments}
\alias{concordances}
\alias{differences}
\alias{show}
\alias{chromosomes,HiCDOCDataSet-method}
\alias{conditions,HiCDOCDataSet-method}
\alias{replicates,HiCDOCDataSet-method}
\alias{resolution,HiCDOCDataSet-method}
\alias{interactions,HiCDOCDataSet-method}
\alias{compartments,HiCDOCDataSet-method}
\alias{differences,HiCDOCDataSet-method}
\alias{concordances,HiCDOCDataSet-method}
\alias{show,HiCDOCDataSet-method}
\title{Methods to manipulate a \code{\link{HiCDOCDataSet}}.}
\format{

















}
\usage{
chromosomes(object)
conditions(object)
replicates(object)
resolution(object)
interactions(object)
compartments(object)
concordances(object)
differences(object)
show(object)


















}
\description{
Retrieve information and results from a \code{\link{HiCDOCDataSet}}.
}
\section{Functions}{
\itemize{
\item \code{chromosomes,HiCDOCDataSet-method}: Retrieves the vector of chromosome names.

\item \code{conditions,HiCDOCDataSet-method}: Retrieves the vector of condition names.

\item \code{replicates,HiCDOCDataSet-method}: Retrieves the vector of replicate names.

\item \code{resolution,HiCDOCDataSet-method}: Retrieves the resolution (span of each position in number of bases).

\item \code{interactions,HiCDOCDataSet-method}: Retrieves a tibble of the interactions.

\item \code{compartments,HiCDOCDataSet-method}: Retrieves a \code{GenomicRange} of the compartment of every position in every
condition.

\item \code{differences,HiCDOCDataSet-method}: Retrieves a \code{GenomicRange} of the significant compartment differences
between conditions, and their p-values.

\item \code{concordances,HiCDOCDataSet-method}: Retrieves a \code{GenomicRange} of the concordance (confidence in assigned
compartment) of every position in every replicate.

\item \code{show,HiCDOCDataSet-method}: Describes the object and its methods.
}}

