% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSet}
\alias{HiCDOCDataSet}
\alias{HiCDOCDatSet}
\alias{HiCDOCDataSet-class}
\title{Infrastructure for HiCDOC data set}
\description{
\code{HiCDOCDataSet} is an S4 class providing the infrastructure (slots)
to store the input data.
}
\details{
\code{HiCDOCDataSet} does this and that...
TODO
}
\section{Slots}{

\describe{
\item{\code{inputMatrix}}{The input matrix}

\item{\code{parameters}}{An named \code{list}. The parameters for the
segmentation methods. See \code{\link{parameters}}.}
}}

