% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HiCDOCDataSetExample}
\alias{HiCDOCDataSetExample}
\title{Example \code{\link{HiCDOCDataSet}} constructor.}
\usage{
HiCDOCDataSetExample()
}
\value{
A \code{\link{HiCDOCDataSet}}.
}
\description{
Constructs a toy \code{\link{HiCDOCDataSet}} from simulated data.
}
\examples{
object <- HiCDOCDataSetExample()

}
