% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSetFromCool}
\alias{HiCDOCDataSetFromCool}
\title{Construct a \code{HiCDOCDataSet} from a cool file.}
\usage{
HiCDOCDataSetFromCool(coolFileNames, replicates, conditions)
}
\arguments{
\item{coolFileNames}{A vector of file names, each one being a .cool file.}

\item{replicates}{The list of the replicate names.}

\item{conditions}{The names of the two conditions.}
}
\value{
\code{HiCDOCDataSet} constructor returns an \code{HiCDOCDataSet}
        object of class S4.
}
\description{
Construct a \code{HiCDOCDataSet} from a cool file.
}
\examples{
basedir <- system.file("extdata", package="HiCDOC")
data    <- read.csv(file.path(basedir, "coolData.csv"))
data
dataSet <- HiCDOCDataSetFromCool(file.path(basedir, data$FileName),
                                 data$Replicate,
                                 data$Condition)

}
