% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HiCDOCDataSetFromHiC}
\alias{HiCDOCDataSetFromHiC}
\title{\code{\link{HiCDOCDataSet}} constructor from HiC files.}
\usage{
HiCDOCDataSetFromHiC(paths, replicates, conditions, resolution)
}
\arguments{
\item{paths}{A vector of paths to \code{.hic} files.}

\item{replicates}{A vector of replicate names repeated along the conditions.}

\item{conditions}{A vector of condition names repeated along the replicates.}

\item{resolution}{The resolution (span of each position in number of bases) to select within
the \code{.hic} files.}
}
\value{
A \code{\link{HiCDOCDataSet}}.
}
\description{
Constructs a \code{\link{HiCDOCDataSet}} from a set of
\code{.hic} files.
}
\examples{
# Retrieve Hi-C files
directory <- system.file("extdata", package = "HiCDOC")
paths <- list.files(directory, '\\\\.hic$')[0:5]
# Specify replicate and condition for each file
# In this case:
# - The first file in the paths vector is replicate 1 of condition a
# - The second file in the paths vector is replicate 2 of condition a
# - ...
# - The last file in the paths vector is replicate x of condition 3
replicates <- c(  1,   2,   1,   2, 'x')
conditions <- c('a', 'a',   2,   2,   3)
# Create HiCDOC object
object <- HiCDOCDataSetFromHiC(
    paths,
    replicates = replicates,
    conditions = conditions,
    resolution = 500000
)

}
