% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSetFromHicPro}
\alias{HiCDOCDataSetFromHicPro}
\title{Construct a \code{HiCDOCDataSet} from hHiC-Pro files}
\usage{
HiCDOCDataSetFromHicPro(
  matrixFileNames,
  bedFileNames,
  replicates,
  conditions,
  cl = NULL
)
}
\arguments{
\item{matrixFileNames}{A vector of file names, each one being a .matrix file.}

\item{bedFileNames}{A vector of file names, each one being a .bed file.}

\item{replicates}{The list of the replicate names.}

\item{conditions}{The names of the two conditions.}
}
\value{
\code{HiCDOCDataSet} constructor returns an \code{HiCDOCDataSet}
        object of class S4.
}
\description{
Construct a \code{HiCDOCDataSet} from hHiC-Pro files
}
