% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{HiCDOCDefaultParameters}
\alias{HiCDOCDefaultParameters}
\title{List of HiCDOC default parameters.}
\format{
An object of class \code{list} of length 7.
}
\usage{
HiCDOCDefaultParameters
}
\description{
List of HiCDOC default parameters.
}
\examples{
exp <- HiCDOCExample()
exp <- HiCDOC(exp, useParameters = HiCDOCDefaultParameters)

}
\keyword{data}
