% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods.R
\docType{class}
\name{HiCDOCExp}
\alias{HiCDOCExp}
\alias{HiCDOCExp-class}
\alias{show,HiCDOCExp-method}
\title{Infrastructure for HiCDOC experiment and differential interaction}
\usage{
HiCDOCExp(matrix = NULL, sampleSize = 1e+05, loessSpan = 0.75)

\S4method{show}{HiCDOCExp}(object)
}
\arguments{
\item{object}{An \code{HiCDOCExp} object.}

\item{inputMatrix}{A matrix with the data.}
}
\value{
\code{HiCDOCExp} constructor returns an \code{HiCDOCExp}
        object of class S4.

The \code{show} method informatively display object contents.
}
\description{
\code{HiCDOC} is an S4 class providing the infrastructure (slots)
to store the input data, methods parameters, intermediate calculations
and results of a differential interaction pipeline
}
\details{
\code{HiCDOCExp} does this and that...
TODO
}
\section{Slots}{

\describe{
\item{\code{inputMatrix}}{The input matrix}

\item{\code{parameters}}{An named \code{list}. The parameters for the
segmentation methods. See \code{\link{parameters}}.}
}}

\examples{
basedir    <- system.file("extdata", package="HiCDOC", mustWork = TRUE)
matrix     <- read.csv(file.path(basedir, "sampleMatrix.tsv"))

srnaExp <- HiCDOCExp(matrix)
srnaExp

}
