% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{interactionMatrix}
\alias{interactionMatrix}
\alias{interactinMatrix,HiCDOCExp-method}
\title{Accessors for the 'interaction' slot of an HiCDOCExp object}
\usage{
\S4method{interactionMatrix}{HiCDOCExp}(object)
}
\arguments{
\item{object}{An \code{HiCDOCExp} object.}
}
\value{
A tibble
}
\description{
The \code{interaction} slot contains the (transformed) interaction profiles.
}
\examples{
exp <- HiCDOCExample()
interactionMatrix(exp)

}
