% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{makeHiCDOCDataSet}
\alias{makeHiCDOCDataSet}
\title{Constructor function for the \code{HiCDOCDataSet} class.}
\usage{
makeHiCDOCDataSet(
  inputPath = NULL,
  interactions = NULL,
  replicates = NULL,
  conditions = NULL,
  binSize = NULL
)
}
\arguments{
\item{inputPath}{One or several file(s) that contain the input matrix/ces.}

\item{interactions}{The interaction matrices.}

\item{replicates}{The list of the replicate names.}

\item{conditions}{The names of the two conditions.}

\item{binSize}{The resolution.}
}
\value{
A \code{HiCDOCDataSet} object.
}
\description{
Constructor function for the \code{HiCDOCDataSet} class.
}
\examples{
basedir <- system.file("extdata", package="HiCDOC", mustWork = TRUE)
matrix  <- file.path(basedir, "sampleMatrix.tsv")
data    <- makeHiCDOCDataSet(inputPath = matrix)
}
