% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{matrixToSparseInteractions}
\alias{matrixToSparseInteractions}
\title{Build the interactions tibble for a chromosome in a condition and replicate.}
\usage{
matrixToSparseInteractions(m, object, chromosomeId, conditionId, replicateId)
}
\arguments{
\item{m}{A matrix.}

\item{object}{A \code{HiCDOCDataSet} object.}

\item{chromosomeId}{A chromosome.}

\item{conditionId}{A condition.}

\item{replicateId}{A replicate.}
}
\value{
An interactions tibble.
}
\description{
Build the interactions tibble for a chromosome in a condition and replicate.
}
