% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{mergeMatrices}
\alias{mergeMatrices}
\title{Merge the matrices which have been parsed separately.
Store the result into a \code{HiCDOCDataSet} object.}
\usage{
mergeMatrices(object, matrices)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object, where the matrices should be
stored.}

\item{matrices}{The different matrices (one per sample).}
}
\value{
object The \code{HiCDOCDataSet} object, where the matrices have been
          added.
}
\description{
Merge the matrices which have been parsed separately.
Store the result into a \code{HiCDOCDataSet} object.
}
