% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalConstructorHelpers.R
\name{modeVector}
\alias{modeVector}
\title{modeVector Extract the mode of vector.}
\usage{
modeVector(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The mode of a vector
}
\description{
modeVector Extract the mode of vector.
}
\examples{
modeVector(c(1, 2, 2, 2, 4))
}
