% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{nbBinsChromosome}
\alias{nbBinsChromosome}
\title{Number of bins of a chromosome, from the interactions matrix}
\usage{
nbBinsChromosome(chromosomeId, interactions, binSize)
}
\arguments{
\item{chromosomeId}{Name of the chromosome}

\item{interactions}{Interactions matrix}

\item{binSize}{Numeric value, resolution}
}
\value{
a numeric value
}
\description{
Number of bins of a chromosome, from the interactions matrix
}
