% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeBiologicalBiases.R
\name{normalizeBiologicalBiases}
\alias{normalizeBiologicalBiases}
\title{Normalize biological biases.}
\usage{
normalizeBiologicalBiases(object)
}
\arguments{
\item{object}{A \code{\link{HiCDOCDataSet}}.}
}
\value{
A \code{\link{HiCDOCDataSet}} with the normalized interactions.
}
\description{
Normalizes biological biases such as GC content and repeated regions. Uses
the Knight-Ruiz balancing algorithm to transform interaction matrices into
doubly stochastic matrices, with sum of rows and sum of columns equal to 1.
}
\examples{
object <- HiCDOCDataSetExample()
object <- filterSparseReplicates(object)
object <- filterWeakPositions(object)
object <- normalizeBiologicalBiases(object)

}
\seealso{
\code{\link{filterSparseReplicates}},
\code{\link{filterWeakPositions}},
\code{\link{HiCDOC}}
}
