% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeBiologicalBiases.R
\name{normalizeBiologicalBiasesChr}
\alias{normalizeBiologicalBiasesChr}
\title{Remove biological biases by normalizing with Knight-Ruiz matrix balancing.}
\usage{
normalizeBiologicalBiasesChr(object, chromosomeId)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{chromosomeId}{The name or number of the chromosome to plot.
If number, will be taken in \code{object@chromosomes[chromosomeId]}}
}
\value{
A \code{HiCDOCDataSet} object, with the normalized matrices.
}
\description{
Remove biological biases by normalizing with Knight-Ruiz matrix balancing.
}
