% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeDistanceEffect.R
\name{normalizeDistanceEffectChr}
\alias{normalizeDistanceEffectChr}
\title{Normalize the distance effect using a loess on the intensity vs distance
to diagonal.}
\usage{
normalizeDistanceEffectChr(object, chromosomeId)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}

\item{chromosomeId}{The name or number of the chromosome to plot.
If number, will be taken in \code{object@chromosomes[chromosomeId]}}
}
\value{
the normalized interaction matrix for this chromosome.
}
\description{
Normalize the distance effect using a loess on the intensity vs distance
to diagonal.
}
