% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods.R
\docType{data}
\name{HiCDOCDefaultParameters}
\alias{HiCDOCDefaultParameters}
\alias{parameters}
\alias{parameters<-}
\alias{parameters<-,HiCDOCDataSet-method}
\title{Print or change the parameters of an HiCDOCDataSet object}
\format{
An object of class \code{list} of length 7.
}
\usage{
HiCDOCDefaultParameters

## S4 replacement method for signature 'HiCDOCDataSet'
parameters(object)

\S4method{parameters}{HiCDOCDataSet}(object) <- value
}
\arguments{
\item{object}{a HiCDOCDataSet object}

\item{value}{a named list with the new parameters values. The names should
be in the \code{HiCDOCDefaultParameters} names.}
}
\description{
parameters Access and print the parameters.
}
\details{
The \code{parameters} slot holds the parameter values
used in an experiment as a named \code{list}. 

Default values exist for parameters, but these can also be supplied 
as input values using the assignment function \code{parameters<-}
or by changing the default parameters in the functions using the 
corresponding parameter, see below. 

\subsection{Parameters in a HiCDOC experiment:}{
   \describe{
       \item{\code{minLengthChr}}{The minimum chromosome
               size (in number of bins), to be kept by the function
               \code{\link{filterSmallChromosomes}}. Default to 100.}
       \item{\code{weakPosThreshold}}{To be kept by the function
               \code{\link{filterWeakPositions}}, the bins (positions)
               of a chromosome must have a mean value greater than
               \code{weakPosThreshold}, on all replicates and all
               conditions. The mean is computed on the row of the
               reconstructed full interaction matrix for 1 chromosome,
               1 condition and 1 replicate. Default to 0.}
       \item{\code{sparseThreshold}}{To be kept by the function
               \code{\link{filterSparseChromosomes}}, the sparsity 
               (percentage of empty cells) of the interactions matrix 
               must be lower than the threshold, on all replicates and 
               all conditions. Default to 0.95.}
      \item{\code{sampleSize}}{The number of bins used when sampling
               all the bins in \code{\link{normalizeDistanceEffect}}.
               Default to 20000.}
      \item{\code{kMeansIterations}}{The maximum number of 2-means
               iterations, used in \code{\link{detectCompartments}} 
               function.}
      \item{\code{kMeansDelta}}{The stop criterion of convergence of
               the 2-means method, used in \code{\link{detectCompartments}} 
               function.}
      \item{\code{kMeansRestarts}}{The maximum number of restarts
               for the 2-means, used in \code{\link{detectCompartments}} 
               function.}
   }
}
}
\section{Functions}{
\itemize{
\item \code{HiCDOCDefaultParameters}: Default parameters for the HiCDOC pipeline

\item \code{parameters<-,HiCDOCDataSet-method}: Change the values of parameters.
}}

\examples{
# All the default parameters
HiCDOCDefaultParameters

object <- HiCDOCExample()

# Default parameters - when object is constructed
parameters(object) 

# Changing a parameter
parameters(object) <- list("minLengthChr" = 50)
parameters(object)
}
\keyword{datasets}
