% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseCoolMatrix}
\alias{parseCoolMatrix}
\title{Parse a single interaction matrix in .cool format.}
\usage{
parseCoolMatrix(fileName)
}
\arguments{
\item{fileName}{The name of the matrix file in HDF5 format.}
}
\value{
object An \code{tibble} storing the (sparse) interaction matrix.
}
\description{
Parse a single interaction matrix in .cool format.
}
