% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseHicMatrix}
\alias{parseHicMatrix}
\title{Parse a single interaction matrix in .hic format.}
\usage{
parseHicMatrix(fileName, resolution = resolution)
}
\arguments{
\item{fileName}{The file name of the .hic file.}

\item{resolution}{The chosen resolution (should be present in the
.hic file).}
}
\value{
object An \code{tibble} storing the (sparse) interaction matrix.
}
\description{
Parse a single interaction matrix in .hic format.
}
