% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseHicPro}
\alias{parseHicPro}
\title{Parse a single interaction matrix in HiC-Pro format}
\usage{
parseHicPro(vectFiles)
}
\arguments{
\item{vectFiles}{2 length vector, of the links to .matrix and .bed files.}
}
\value{
object An \code{tibble} storing the (sparse) interaction matrix.
}
\description{
Parse a single interaction matrix in HiC-Pro format
}
