% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseInteractionMatrix3Columns}
\alias{parseInteractionMatrix3Columns}
\title{Read interaction matrices in tabular 3-column format.}
\usage{
parseInteractionMatrix3Columns(object)
}
\arguments{
\item{object}{An \code{HiCDOCExp} object.}
}
\value{
object An \code{HiCDOCExp} object.
}
\description{
This function parses the tsv files provided in
\code{object@inputMatrixPath}.
}
\examples{
basedir <-
  system.file("extdata", package = "HiCDOC", mustWork = TRUE)
matrix  <- file.path(basedir, "sample.tsv")
data   <- makeHiCDOCDataSet(inputPath = matrix)
object <- parseInteractionMatrix3Columns(data)
}
