% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseInteractionMatrixHicPro}
\alias{parseInteractionMatrixHicPro}
\title{Read interaction matrices in Hic-Pro format.}
\usage{
parseInteractionMatrixHicPro(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
object An \code{HiCDOCDataSet} object.
}
\description{
This function parses the HiC-Pro files provided in
\code{object@inputPath}.
}
