% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseInteractionMatrixMCool}
\alias{parseInteractionMatrixMCool}
\title{Read interaction matrices in .mcool format.}
\usage{
parseInteractionMatrixMCool(object, resolution)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}

\item{resolution}{The chosen resolution, in base pairs.}
}
\value{
object    An \code{HiCDOCDataSet} object.
}
\description{
This function parses the .mcool files provided in
\code{object@inputMatrixPath} at a given resolution.
}
