% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseMCoolMatrix}
\alias{parseMCoolMatrix}
\title{Parse a single interaction matrix in .mcool format.}
\usage{
parseMCoolMatrix(fileName, resolution = resolution)
}
\arguments{
\item{fileName}{The name of the matrix file in HDF5 format.}

\item{resolution}{The chosen resolution (in bp)}
}
\value{
object An \code{tibble} storing the (sparse) interaction matrix.
}
\description{
Parse a single interaction matrix in .mcool format.
}
