% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompartmentChanges.R
\name{plotAllCompartmentChanges}
\alias{plotAllCompartmentChanges}
\title{Plot the changes of compartments, for all chromosomes}
\usage{
plotAllCompartmentChanges(
  object,
  padjThreshold = 0.05,
  xlim = NULL,
  points = FALSE
)
}
\arguments{
\item{object}{a HicDOCExp object, on which \code{\link{detectCompartments}} have run}

\item{padjThreshold}{threshold for the adjusted p-value to show significant changes. Default to 0.05}

\item{xlim}{numerical vector of length 2. Giving the limits on the x-axis (position) to show. Default to NULL}

\item{points}{boolean (default to FALSE). If TRUE, points are represented on the concordance lines.}
}
\value{
A list of \code{ggplot} objects, one for each chromosome.
}
\description{
Plot the changes of compartments, for all chromosomes
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotCompartmentsChanges(object)
}
