% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCentroids}
\alias{plotCentroids}
\title{Plot the centroid distributions along the genomic positions for a given chromosome.}
\usage{
plotCentroids(object, chromosomeId)
}
\arguments{
\item{chromosomeId}{Name or number of the chromosome, like in object@chromosome}

\item{objet}{an \code{HiCDOCExp} object}
}
\value{
A \code{ggplot} object
}
\description{
Plot the centroid distributions along the genomic positions for a given chromosome.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotCentroids(object, 1)
}
