% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompartmentChanges.R
\name{plotCompartmentChanges}
\alias{plotCompartmentChanges}
\title{Run plotCompartments() and plotConcordances() and assemble them on the
same plot}
\usage{
plotCompartmentChanges(
  object,
  chromosome,
  threshold = 0.05,
  xlim = NULL,
  points = FALSE
)
}
\arguments{
\item{object}{An HiCDOCDataSet object, after a detectCompartments() run}

\item{threshold}{Significance threshold for the changes. Default to 0.05.}

\item{xlim}{A numeric-value pair, indicating the interval of positions
to represent.
Default to NULL = all positions.}

\item{points}{Logical (default to FALSE). If TRUE, points will be added
on the concordance lines.}

\item{chromosomeId}{Name or number of the chromosome, like in
object@chromosome}
}
\value{
A ggplot object.
}
\description{
Run plotCompartments() and plotConcordances() and assemble them on the
same plot
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotCompartmentChanges(object, 1)
}
