% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompartmentChanges.R
\name{plotCompartments}
\alias{plotCompartments}
\title{Plot the A & B compartments}
\usage{
plotCompartments(object, chromosomeId, xlim = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object on which
\code{detectCompartments()} has run.}

\item{chromosomeId}{The name or number of the chromosome to plot.
If number, will be taken in \code{object@chromosomes[chromosomeId]}}

\item{xlim}{A numeric-value pair, indicating the interval of positions
to represent.
Default to NULL = all positions.}
}
\value{
A ggplot.
}
\description{
Plot the A and B compartments after \code{detectCompartments()} on
a HiCDOCDataSet object, for a chromosome.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotCompartments(object, 1)
}
