% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConcordanceDistribution.R
\name{plotConcordanceDistribution}
\alias{plotConcordanceDistribution}
\title{Plot the concordance, i.e. the relative distance of the genomic positions
with respect to the centroids.}
\usage{
plotConcordanceDistribution(object)
}
\arguments{
\item{object}{an \code{HiCDOCDataSet} object}
}
\value{
A list of \code{ggplot}, one for each chromosome.
}
\description{
Plot the concordance, i.e. the relative distance of the genomic positions
with respect to the centroids.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotConcordanceDistribution(object)
}
