% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotConcordances}
\alias{plotConcordances}
\title{Plot the concordance, i.e. the relative distance of the genomic positions with respect to the centroids.}
\usage{
plotConcordances(object)
}
\arguments{
\item{objet}{an \code{HiCDOCExp} object}
}
\value{
A list of \code{ggplot}, one for each chromosome.
}
\description{
Plot the concordance, i.e. the relative distance of the genomic positions with respect to the centroids.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotConcordances(object)
}
