% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotInteractionMatrix}
\alias{plotInteractionMatrix}
\title{Plot the interaction matrices (as heatmaps).}
\usage{
plotInteractionMatrix(object, log)
}
\arguments{
\item{log}{a logical scalar: should the color be in log scale}

\item{objet}{an \code{HiCDOCExp} object}
}
\value{
A list of \code{ggplot}, one for each chromosome.
}
\description{
Plot the interaction matrices (as heatmaps).
}
\examples{
object <- HiCDOCExample()
p <- plotInteractionMatrix(object, log = TRUE)
}
