% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelfInteractionRatios.R
\name{plotSelfInteractionRatios}
\alias{plotSelfInteractionRatios}
\title{Plot boxplots of the diagonal ratios in each compartment,
for a given chromosome.}
\usage{
plotSelfInteractionRatios(object, chromosome)
}
\arguments{
\item{object}{an \code{HiCDOCDataSet} object}

\item{chromosome}{character or numeric value, name or number of chromosome}
}
\value{
A list of \code{ggplot}, one for each chromosome.
}
\description{
Plot boxplots of the diagonal ratios in each compartment,
for a given chromosome.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotSelfInteractionRatios(object, 1)
}
