% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{positions}
\alias{positions}
\alias{positions,HiCDOCDataSet-method}
\title{Accessors for the 'positions' slot of an HiCDOCDataSet object}
\usage{
\S4method{positions}{HiCDOCDataSet}(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
A tibble
}
\description{
The \code{positions} slot contains the positions of the bins, by chromosome
}
\examples{
exp <- HiCDOCExample()
positions(exp)

}
