% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{predictABCompartments}
\alias{predictABCompartments}
\title{Use difference between diagonal and off-diagonal interactions to determine
which clusters correspond to compartments A and B.}
\usage{
predictABCompartments(object)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}
}
\value{
A \code{HiCDOCExp} object, with A and B labels replacing cluster
numbers in centroids, compartments, distances and concordances.
}
\description{
Use difference between diagonal and off-diagonal interactions to determine
which clusters correspond to compartments A and B.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- clusterize(object)
object <- predictABCompartments(object)
}
