% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{reduceHiCDOCDataSet}
\alias{reduceHiCDOCDataSet}
\title{reduce an HiCDOCDataSet}
\usage{
reduceHiCDOCDataSet(
  object,
  chromosomes = NULL,
  conditions = NULL,
  replicates = NULL
)
}
\arguments{
\item{object}{and HiCDOCDataSet object}

\item{chromosome}{default to NULL, chromohomes ID to keep}

\item{condition}{default to NULL, conditions ID to keep}

\item{replicate}{default to NULL, replicate ID to keep}
}
\value{
a HiCDOCDataSET object, reduced by keeping only the chromosomes, 
conditions or replicates given in parameters
}
\description{
reduce an HiCDOCDataSet
}
\examples{
object <- HiCDOCExample()
objectReduced <-reduceHiCDOCDataSet(object, chromosome = "17", replicate = "1")
}
