% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpersConstructor.R
\name{reformatInteractions}
\alias{reformatInteractions}
\title{The function takes an interaction matrix and performs some correction
- the matrix is putted in upper format only (the values in the lower side
will be reaffected in the upper side)
- the value column is transformed in numeric
- the duplicated combinations of positions are aggregated by mean}
\usage{
reformatInteractions(interactions)
}
\arguments{
\item{interactions}{the slot interactions of a HiCDOCDataSet object.}
}
\value{
a tibble, of the interactions corrected
}
\description{
The function takes an interaction matrix and performs some correction
- the matrix is putted in upper format only (the values in the lower side
will be reaffected in the upper side)
- the value column is transformed in numeric
- the duplicated combinations of positions are aggregated by mean
}
