% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpersConstructor.R
\name{replacePositionsByBins}
\alias{replacePositionsByBins}
\title{The function takes an HiCDOC object and replace the positions by the
corresponding bins in the interaction matrix}
\usage{
replacePositionsByBins(object)
}
\arguments{
\item{object}{a HiCDOCDataSet object}
}
\value{
a tibble, of the modified interactions matrix.
}
\description{
The function takes an HiCDOC object and replace the positions by the
corresponding bins in the interaction matrix
}
