% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{runHiCDOC}
\alias{runHiCDOC}
\title{Launch the HiCDOC pipeline.}
\usage{
runHiCDOC(object, parallel = FALSE)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{parallel}{Logical, default to FALSE. Should parallel computation 
should be used ?}
}
\value{
Returns an \code{HiCDOCDataSet} object.
}
\description{
This functions allow to run all the recommended steps to run the HiCDOC 
pipeline.
}
\details{
\subsection{The pipeline have 7 steps :}{
   \describe{
       \item{3 filters to clean the data:}{
       \itemize{
           \item{\code{\link{filterSmallChromosomes}}}{
           to filter the too small chromosomes}
           \item{\code{\link{filterWeakPositions}}}{
           to filter the "weak" positions of a chromosome}
           \item{\code{\link{filterSparseChromosomes}}}{
           to filter the too sparse interactions matrix
           by chromosome}
       }
     }
   \item{3 normalization steps to normalize the data:}{
   \itemize{
     \item{\code{\link{normalizeTechnicalBiases}}}{
     to normalize the data in case of tehnical bias}
     \item{\code{\link{normalizeBiologicalBiases}}}{
     to normalize the data in case of biological bias}
     \item{\code{\link{normalizeDistanceEffect}}}{
     to normalize the distance effect}
   }
   }
   \item{the detection of compartments}{
\itemize{
     \item{\code{\link{detectCompartments}}}{
     to run the detection of compartments and sinificant differences}
   }
 }
   }
}
}
\examples{
object <- HiCDOCExample()
object <- runHiCDOC(object)
}
