% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,HiCDOCDataSet-method}
\title{Show the components of a HiCDOC objects}
\usage{
\S4method{show}{HiCDOCDataSet}(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
The \code{show} method informatively display object contents.
}
\description{
This function informatively display object contents.
}
