% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sparseInteractionsToFullInteractions}
\alias{sparseInteractionsToFullInteractions}
\title{Build the full interactions tibble filled with zeros for a chromosome in a
condition and replicate.}
\usage{
sparseInteractionsToFullInteractions(
  object,
  chromosomeId,
  conditionId,
  replicateId,
  filter = TRUE
)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}

\item{chromosomeId}{A chromosome.}

\item{conditionId}{A condition.}

\item{replicateId}{A replicate.}

\item{filter}{Whether to remove weak positions. Defaults to TRUE.}
}
\value{
An interactions tibble.
}
\description{
Build the full interactions tibble filled with zeros for a chromosome in a
condition and replicate.
}
