% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sparseInteractionsToMatrix}
\alias{sparseInteractionsToMatrix}
\title{Build the interaction matrix for a chromosome in a condition and replicate.}
\usage{
sparseInteractionsToMatrix(
  object,
  chromosomeId,
  conditionId,
  replicateId,
  filter = FALSE
)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{chromosomeId}{A chromosome.}

\item{conditionId}{A condition.}

\item{replicateId}{A replicate.}

\item{filter}{Shrink the matrix by removing weak rows/columns.}
}
\value{
A matrix.
}
\description{
Build the interaction matrix for a chromosome in a condition and replicate.
}
