% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSparseChromosomes.R
\name{sparsityChromosome}
\alias{sparsityChromosome}
\title{sparsityChromosome
Compute the maximum of sparsity of a given chromosome}
\usage{
sparsityChromosome(object, chromosomeId)
}
\arguments{
\item{object}{a HiCDOCDataSet object}

\item{chromosomeId}{the name of the chromosome}
}
\value{
a one line tibble
}
\description{
sparsityChromosome
Compute the maximum of sparsity of a given chromosome
}
