% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validParameters.R
\name{testchromosome}
\alias{testchromosome}
\title{Return valid chromosome as string}
\usage{
testchromosome(object, chrId)
}
\arguments{
\item{object}{An HiCDOC object}

\item{chrId}{Numerical or character value}
}
\value{

}
\description{
Test if the chromosome entred exists. 
If a numerical value is verified, and exists, return the 
corresponding character in object@chromosome
}
