% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validParameters.R
\name{testCondition}
\alias{testCondition}
\title{Test the existence of a given condition in a HiCDOC object}
\usage{
testCondition(object, conditionId)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{conditionId}{A character or numeric value}
}
\value{
The condition name, or an error.
}
\description{
Test the existence of a given condition in a HiCDOC object
}
