% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validParameters.R
\name{testNbCores}
\alias{testNbCores}
\title{testNbCores
Test if the nbCores parameter is correct, for parallel computations.}
\usage{
testNbCores(nbCores)
}
\arguments{
\item{nbCores}{integer. The number of cores to use in parallel::mcmapply.}
}
\value{
integer
}
\description{
testNbCores
Test if the nbCores parameter is correct, for parallel computations.
}
