% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validParameters.R
\name{testSlotsHiCDOC}
\alias{testSlotsHiCDOC}
\title{Test the existence of slots in HiCDOCDataSet object}
\usage{
testSlotsHiCDOC(object, slots = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{slots}{Character vector, names of slots to verify. Default to NULL.
If NULL, check only for the class of \code{object}}
}
\value{
An error if the object is not a HiCDOCDataSet object or a slot is
missing.
}
\description{
Test the existence of slots in HiCDOCDataSet object
}
