% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validParameters.R
\name{testSlotsHiCDOCExp}
\alias{testSlotsHiCDOCExp}
\title{Test the existence of slots in HiCDOCExp object}
\usage{
testSlotsHiCDOCExp(object, slots = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}

\item{slots}{Names of slots to verify.}
}
\value{
An error if the object is not a HiCDOCExp object or a slot is
missing.
}
\description{
Test the existence of slots in HiCDOCExp object
}
