% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validParameters.R
\name{testValidId}
\alias{testValidId}
\title{Test the existence of a set of identifier}
\usage{
testValidId(object, id, what = "chromosomes")
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{id}{The identifiers to test (string or integer)}

\item{what}{In what slot should we test the presence of the identifiers ? 
One of "chromosome", "condition", "replicate"}
}
\value{
The chromosome name or an error.
}
\description{
Test the existence of a set of identifier
}
