% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompartmentChanges.R
\name{textsignif}
\alias{textsignif}
\title{Return the text indicating what represent the gray areas on
plotConcordance()}
\usage{
textsignif(differences, padjThreshold)
}
\arguments{
\item{differences}{difference, from object@differences after
detectCompartments() function}

\item{padjThreshold}{threshold for the significance}
}
\value{
A character vector of length 1
}
\description{
Return the text indicating what represent the gray areas on
plotConcordance()
}
\keyword{internal}
