% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{tieCentroids}
\alias{tieCentroids}
\title{Assign correct cluster labels by comparing centroids across conditions.}
\usage{
tieCentroids(object)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}
}
\value{
A \code{HiCDOCDataSet} object, with diagonalRatios, and with corrected
cluster labels in centroids, compartments, distances and concordances.
}
\description{
Assign correct cluster labels by comparing centroids across conditions.
}
