% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validParameters.R
\name{checkParameters}
\alias{checkParameters}
\title{Check object@parameters, return default if NULL}
\usage{
checkParameters(parameters)
}
\arguments{
\item{parameters}{list of parameters}
}
\value{
list of updated parameters, default from HiCDOCDefaultParameters()
if null.
}
\description{
Check object@parameters, return default if NULL
}
