% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{chromosomes}
\alias{chromosomes}
\alias{chromosomes,HiCDOCExp-method}
\title{Accessors for the 'chromosomes' slot of an HiCDOCExp object}
\usage{
\S4method{chromosomes}{HiCDOCExp}(object)
}
\arguments{
\item{object}{An \code{HiCDOCExp} object.}
}
\value{
A character vector
}
\description{
The \code{chromosomes} slot contains the names of the chromosomes, eventually
filtred after \code{filterSmallChromosomes()}
}
\examples{
exp <- HiCDOCExample()
chromosomes(exp)

}
