% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{clusterize}
\alias{clusterize}
\title{Segregate genomic positions into 2 clusters using constrained k-means.}
\usage{
clusterize(object)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}
}
\value{
A \code{HiCDOCExp} object, with:
- centroid (vector) for each cluster
- compartment (cluster number) for each genomic position
- distances to centroids (float) for each genomic position in each replicate
- concordance (float in [-1, 1]) for each genomic position in each replicate
}
\description{
Segregate genomic positions into 2 clusters using constrained k-means.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- clusterize(object)
}
