% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{computePValues}
\alias{computePValues}
\title{Get p-values for genomic positions whose assigned compartment switches
between two conditions:
1. For each pair of replicates in different conditions, for each genomic
   position, compute the absolute difference between its concordances.
2. For each pair of conditions, for each genomic position, compute the
   median of its concordance differences.
3. For each pair of conditions, for each genomic position whose assigned
   compartment switches, rank its median against the empirical cumulative
   distribution of medians for all non-switching positions in that condition
   pair. Adjust the resulting p-value with the Benjamini–Hochberg procedure.}
\usage{
computePValues(object)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}
}
\value{
A \code{HiCDOCExp} object, with differences and their p-values.
}
\description{
Get p-values for genomic positions whose assigned compartment switches
between two conditions:
1. For each pair of replicates in different conditions, for each genomic
   position, compute the absolute difference between its concordances.
2. For each pair of conditions, for each genomic position, compute the
   median of its concordance differences.
3. For each pair of conditions, for each genomic position whose assigned
   compartment switches, rank its median against the empirical cumulative
   distribution of medians for all non-switching positions in that condition
   pair. Adjust the resulting p-value with the Benjamini–Hochberg procedure.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- clusterize(object)
object <- predictABCompartments(object)
object <- computePValues(object)
}
