% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{detectCompartments}
\alias{detectCompartments}
\title{Detect the compartments for each position of the genome.
Call the differential binding.}
\usage{
detectCompartments(object)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}
}
\value{
A \code{HiCDOCExp} object, with the compartment calling.
}
\description{
Detect the compartments for each position of the genome.
Call the differential binding.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
}
