% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{detectCompartments}
\alias{detectCompartments}
\title{Detect compartments for each genomic position in each condition, and compute
p-values for compartment differences between conditions.}
\usage{
detectCompartments(object, parallel = FALSE, nbCores = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{parallel}{ogical. Should parallel processing be used? Default to FALSE.}

\item{nbCores}{Integer, the number of cores to use for parallelization.}
}
\value{
A \code{HiCDOCDataSet} object, with centroids, compartments, distances,
concordances and differences.
}
\description{
Detect compartments for each genomic position in each condition, and compute
p-values for compartment differences between conditions.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
}
