% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{detectCompartments}
\alias{detectCompartments}
\title{Detect compartments for each genomic position in each condition, and compute
p-values for compartment differences between conditions.}
\usage{
detectCompartments(object)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}
}
\value{
A \code{HiCDOCExp} object, with centroids, compartments, distances,
concordances and differences.
}
\description{
Detect compartments for each genomic position in each condition, and compute
p-values for compartment differences between conditions.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
}
