% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{diagonalRatios}
\alias{diagonalRatios}
\title{Compute diagonalRatios}
\usage{
diagonalRatios(object, chromosomeId, conditionId, replicateId)
}
\arguments{
\item{object}{an HiCDOCDataSet object}

\item{chromosomeId}{A character or numeric value.
Name or number of the chromosome}

\item{conditionId}{A character or numeric value.
Name or number of the condition}

\item{replicateId}{A character or numeric value.
Name or number of the replicate}
}
\value{
a tibble
}
\description{
Compute diagonalRatios
}
