% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{differences}
\alias{differences}
\alias{differences,HiCDOCDataSet-method}
\title{Extracts differentially interacting regions of an HiCDOCDataSet object}
\usage{
\S4method{differences}{HiCDOCDataSet}(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
A \code{GenomicRanges} object of the selected differentially
interacting regions.
}
\description{
This function extracts the differentially interacting regions from
\code{\link{HiCDOCDataSet}}.
}
\examples{
exp <- HiCDOCExample()
exp <- HiCDOC(exp)
differences(exp)
}
