% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{differences}
\alias{differences}
\alias{differences,HiCDOCExp-method}
\title{Extracts differentially interacting regions of an HiCDOCExp object}
\usage{
\S4method{differences}{HiCDOCExp}(object, pvalue = 1)
}
\arguments{
\item{object}{An \code{HiCDOCExp} object.}

\item{pvalue}{Numeric cutoff value for adjusted p-values. Only regions with
adjusted p-values equal or lower than specified are returned.
Default to 1, all regions are returned.}
}
\value{
A \code{GenomicRanges} object of the selected differentially
interacting regions.
}
\description{
This function extracts the differentially interacting regions from
\code{\link{HiCDOCExp}}.
}
\examples{
exp <- HiCDOCExample()
exp <- HiCDOC(exp)
differences(exp, pvalue = 0.05)

}
