% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{distanceRatio}
\alias{distanceRatio}
\title{Compute the log ratio of the distance of a position to each centroid.}
\usage{
distanceRatio(x, centroids, eps = 1e-10)
}
\arguments{
\item{x}{A vector of a genomic position.}

\item{centroids}{A list of 2 vectors.}

\item{eps}{A small epsilon to avoid log(0).}
}
\value{
A float number.
}
\description{
Compute the log ratio of the distance of a position to each centroid.
}
\examples{
distanceRatio(c(1.5,2,3.8), list('1'=c(0.5,1.1,0.4), '2'=c(2.3,4.8,1)))
}
