% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Compute the euclidean distance between two vectors.}
\usage{
euclideanDistance(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A float number.
}
\description{
Compute the euclidean distance between two vectors.
}
