% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleHiCDOCDataSetProcessed}
\alias{exampleHiCDOCDataSetProcessed}
\title{Example HiCDOCDataSet, filtered and normalized.}
\format{
S4 HiCDOCDataSet object with the following characteristics:
\describe{
  \item{chromosomes}{4 chromosomes: X, Y, Z}
  \item{conditions}{3 conditions: 1, 2, 3}
  \item{replicates}{3 replicates: R1, R2, R3}
  \item{binSize}{A resolution of 137 bases}
}
}
\usage{
exampleHiCDOCDataSetProcessed
}
\value{
A \code{\link{HiCDOCDataSet}}, already filtered and normalized.
}
\description{
A S4 HiCDOCDataSet object with 3 chromosomes, 3 conditions and 3 replicates.
}
\keyword{datasets}
