% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSmallChromosomes.R
\name{filterSmallChromosomes}
\alias{filterSmallChromosomes}
\title{Suppress the small chromosomes}
\usage{
filterSmallChromosomes(object, minLength = NULL)
}
\arguments{
\item{object}{A HiCDOCDataSet object}

\item{minLength}{Numeric value, default to
HiCDOCDefaultParameters$minLengthChr. The minimum chromosome
size (in number of bins), to be kept.}
}
\value{
A HiCDOCDataSet object
}
\description{
The function return an HiCDOCDataSet object, with only the big chromosomes.
The big chromosomes must have a totalBins >= minLength.
The interactions are reduced to keep only those corresponding
to the remaining chromosomes.
}
\examples{
object <- HiCDOCExample()
chromosomes(object)
object <- filterSmallChromosomes(object)
chromosomes(object)
}
