% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSmallChromosomes.R
\name{filterSmallChromosomes}
\alias{filterSmallChromosomes}
\title{Suppress the small chromosomes}
\usage{
filterSmallChromosomes(object)
}
\arguments{
\item{object}{A HiCDOC object}
}
\value{
A HiCDOC object
}
\description{
The function return an HiCDOCExp object, with only the big chromosomes.
The big chromosomes must have a totalBins > object@minLength.
The interactions are reduced to keep only those corresponding
to the remaining chromosomes
}
\examples{
object <- HiCDOCExample()
object@chromosomes
object <- filterSmallChromosomes(object)
object@chromosomes
}
