% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSmallChromosomes.R
\name{filterSmallChromosomes}
\alias{filterSmallChromosomes}
\title{Suppress the small chromosomes}
\usage{
filterSmallChromosomes(object, minLength = 100)
}
\arguments{
\item{object}{A HiCDOC object}

\item{minLength}{Numeric value, default to 100. The minimum chromosome 
size (in number of bins), to be kept.}
}
\value{
A HiCDOC object
}
\description{
The function return an HiCDOCExp object, with only the big chromosomes.
The big chromosomes must have a totalBins >= minLength.
The interactions are reduced to keep only those corresponding
to the remaining chromosomes.
}
\examples{
object <- HiCDOCExample()
object@chromosomes
object <- filterSmallChromosomes(object)
object@chromosomes
}
