% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterWeakPositions.R
\name{filterWeakPositions}
\alias{filterWeakPositions}
\title{Remove the weak bins of an HiCDOCDataSet object}
\usage{
filterWeakPositions(object, threshold = NULL)
}
\arguments{
\item{object}{A HiCDOCDataSet object}

\item{threshold}{Numeric. Threshold to consider bins as 'weaks'. If NULL,
default to the first not NULL of \code{object$weakPosThreshold} and
\code{HiCDOCDefaultParameters$weakPosThreshold}.}
}
\value{
A \code{HiCDOCDataSet} object with a reduced \code{interactions}
slot, and the weak bins identified by chromosomes in \code{object@weakBins}.
}
\description{
The function indentifies and remove the weak bins of an HiCDOCDataSet object,
chromosome by chromosome.
To be kept, the bins (positions)of a chromosome must have a mean value
greater than \code{threshold}, on all replicates and all conditions.
The mean is computed on the row of the reconstructed full interaction
matrix for 1 chromosome, 1 condition and 1 replicate. The weak bins will
be discarded, and added in object@weakBins.
}
\examples{
exp <- HiCDOCExample()
exp <- filterWeakPositions(exp)
}
\seealso{
\code{\link[HiCDOC]{filterSmallChromosomes}}, 
\code{\link[HiCDOC]{filterSparseChromosomes}} and 
\code{\link[HiCDOC]{HiCDOC}} for the recommended pipeline.
}
