% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterWeakPositions.R
\name{filterWeakPositions}
\alias{filterWeakPositions}
\title{Remove the weak bins of an HiCDOC object}
\usage{
filterWeakPositions(object, threshold = 0)
}
\arguments{
\item{object}{A HiCDOC object}

\item{threshold}{Numeric, default to 0.}
}
\value{
A \code{HicDOCExp} object with a reduced \code{interactions} slot, 
and the weak bins identified by chromosomes in \code{object@weakBins}.
}
\description{
The function indentifies and remove the weak bins of an HiCDOC object, 
chromosome by chromosome. 
To be kept, the bins (positions) must have a mean value greater than 
\code{threshold}, on all replicates and all conditions. 
The mean is computed on the row of the reconstructed full interaction 
matrix for 1 chromosome, 1 condition and 1 replicate. The weak bins will
be discarded, and added in object@weakBins.
}
\examples{
exp <- HiCDOCExample()
exp <- filterWeakPositions(exp)
}
