% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCDOC-package.R
\docType{package}
\name{HiCDOC-package}
\alias{HiCDOC-package}
\alias{_PACKAGE}
\title{A/B compartment detection and differential analysis}
\description{
HiCDOC normalizes intrachromosomal Hi-C matrices, uses unsupervised learning
    to predict A/B compartments from multiple replicates, and detects
    significant compartment changes between experiment conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mzytnicki/HiCDOC}
  \item Report bugs at \url{https://github.com/mzytnicki/HiCDOC/issues}
}

}
\author{
\strong{Maintainer}: Zytnicki Matthias \email{matthias.zytnicki@inrae.fr}

Authors:
\itemize{
  \item Kurylo Cyril \email{cyril.kurylo@inrae.fr}
  \item Foissac Sylvain \email{sylvain.foissac@inrae.fr}
  \item Maign<U+00E9> <U+00C9>lise \email{elise.maigne@inrae.fr}
}

}
