% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCDOC-package.R
\docType{package}
\name{HiCDOC-package}
\alias{HiCDOC}
\alias{HiCDOC-package}
\title{Find compartments with significantly differential binding using
replicated Hi-C data}
\description{
Find compartments with significantly differential binding using 
    replicated Hi-C data.
    bla bla bla
    bla bla bla from the DESCRIPTION file
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mzytnicki/HiCDOC}
  \item Report bugs at \url{https://github.com/mzytnicki/HiCDOC/issues}
}

}
\author{
\strong{Maintainer}: Zytnicki Matthias \email{matthias.zytnicki@inrae.fr}

Authors:
\itemize{
  \item Kurylo Cyril \email{Cyril.Kurylo@inrae.fr}
  \item Foissac Sylvain \email{sylvain.foissac@inrae.fr}
  \item Maigné Élise \email{elise.maigne@inrae.fr}
}

}
