% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCDOC-package.R
\docType{package}
\name{HiCDOC-package}
\alias{HiCDOC-package}
\alias{_PACKAGE}
\alias{HiCDOC}
\title{HiCDOC: Robust differential compartment detection}
\description{
Find compartments with significantly differential binding using 
    replicated Hi-C data.
    bla bla bla
    bla bla bla
}
\author{
\strong{Maintainer}: Zytnicki Matthias \email{matthias.zytnicki@inrae.fr}

Authors:
\itemize{
  \item Kurylo Cyril \email{Cyril.Kurylo@inrae.fr}
  \item Foissac Sylvain \email{sylvain.foissac@inrae.fr}
  \item Maigné Élise \email{elise.maigne@inrae.fr}
}

}
\keyword{internal}
