% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCDOC-package.R
\docType{package}
\name{HiCDOC-package}
\alias{HiCDOC-package}
\alias{_PACKAGE}
\title{A/B compartment detection and differential analysis}
\description{
HiCDOC normalizes intrachromosomal Hi-C matrices, uses unsupervised learning to predict A/B compartments from multiple replicates, and detects significant compartment changes between experiment conditions. It provides a collection of functions assembled into a pipeline to filter and normalize the data, predict the compartments and visualize the results. It accepts several type of data: tabular `.tsv` files, Cooler `.cool` or `.mcool` files, Juicer `.hic` files or HiC-Pro `.matrix` and `.bed` files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mzytnicki/HiCDOC}
  \item Report bugs at \url{https://github.com/mzytnicki/HiCDOC/issues}
}

}
\author{
\strong{Maintainer}: Maigné Élise \email{elise.maigne@inrae.fr}

Authors:
\itemize{
  \item Kurylo Cyril \email{cyril.kurylo@inrae.fr}
  \item Zytnicki Matthias \email{matthias.zytnicki@inrae.fr}
  \item Foissac Sylvain \email{sylvain.foissac@inrae.fr}
}

}
