% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/HiCDOC-package.R
\docType{package}
\name{HiCDOC}
\alias{HiCDOC}
\alias{_PACKAGE}
\alias{HiCDOC-package}
\title{Find compartments with significantly differential binding using 
replicated Hi-C data}
\usage{
HiCDOC(object, parallel = FALSE) # to run the entire HiCDOC pipeline
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{parallel}{Logical, default to FALSE. Should parallel computation 
should be used ?}
}
\value{
Returns an \code{HiCDOCDataSet} object.
}
\description{
To learn more about HiCDOC, start with the vignette:
\code{browseVignettes(package = "HiCDOC")}.

The HiCDOC function allow to run the HiCDOC pipeline, on a HiCDOCDataSet 
object.

Find compartments with significantly differential binding using 
    replicated Hi-C data.
    bla bla bla
    bla bla bla from the DESCRIPTION file
}
\details{
\subsection{HiCDOCDataSet object}{
    A \code{HiCDOCDataSet} object can be created from different object : 
    \itemize{
        \item{matrices: }{see \code{\link{HiCDOCDataSetFromSparseMatrix}}},
        \item{cool or mcool data: }{see \code{\link{HiCDOCDataSetFromCool}}},
        \item{hic data: }{see \code{\link{HiCDOCDataSetFromHic}}} and 
        \item{HicPro data: }{see \code{\link{HiCDOCDataSetFromHicPro}}}
    }
}
\subsection{The HiCDOCDataSet pipeline have 7 steps :}{
   \describe{
       \item{3 filters to clean the data:}{
       \itemize{
           \item{\code{\link{filterSmallChromosomes}}}{
           to filter the too small chromosomes}
           \item{\code{\link{filterWeakPositions}}}{
           to filter the "weak" positions of a chromosome}
           \item{\code{\link{filterSparseChromosomes}}}{
           to filter the too sparse interactions matrix
           by chromosome}
       }
     }
   \item{3 normalization steps to normalize the data:}{
   \itemize{
     \item{\code{\link{normalizeTechnicalBiases}}}{
     to normalize the data in case of tehnical bias}
     \item{\code{\link{normalizeBiologicalBiases}}}{
     to normalize the data in case of biological bias}
     \item{\code{\link{normalizeDistanceEffect}}}{
     to normalize the distance effect}
   }
   }
   \item{the detection of compartments}{
\itemize{
     \item{\code{\link{detectCompartments}}}{
     to run the detection of compartments and compute 
     significant differences}
   }
 }
   }
}
}
\examples{
object <- HiCDOCExample()
object <- HiCDOC(object)

# This is equivalent of
\dontrun{
parallel = FALSE
object <- filterSmallChromosomes(object)
object <- filterSparseChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object, parallel = parallel)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object, parallel = parallel)
}
}
\seealso{
\code{\link{HiCDOCDataSet}}

Useful links:
\itemize{
  \item \url{https://github.com/mzytnicki/HiCDOC}
  \item Report bugs at \url{https://github.com/mzytnicki/HiCDOC/issues}
}

}
\author{
\strong{Maintainer}: Zytnicki Matthias \email{matthias.zytnicki@inrae.fr}

Authors:
\itemize{
  \item Kurylo Cyril \email{Cyril.Kurylo@inrae.fr}
  \item Foissac Sylvain \email{sylvain.foissac@inrae.fr}
  \item Maigné Élise \email{elise.maigne@inrae.fr}
}

}
