% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/HiCDOC-package.R
\docType{package}
\name{HiCDOC}
\alias{HiCDOC}
\alias{HiCDOC-package}
\title{Main method.  Start the pipeline with default parameters.}
\usage{
HiCDOC(object, parallel = FALSE)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{parallel}{Logical, default to FALSE. Should parallel computation should be used ?}
}
\value{
Returns an \code{HiCDOCDataSet} object.
}
\description{
\code{HiCODC} is a package...
TODO
}
\examples{
object <- HiCDOCExample()
object <- HiCDOC(object)
# TODO
}
\author{
Cyril Kurylo and Matthias Zytnicki and Sylvain Foissac and
Élise Maigné
}
\keyword{package}
