% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCDOC-package.R
\docType{package}
\name{HiCDOC}
\alias{HiCDOC}
\alias{HiCDOC-package}
\title{Title...}
\description{
\code{HiCODC} is a package...
TODO
}
\examples{

}
\author{
Cyril Kurylo and Matthias Zytnicki and Sylvain Foissac
}
\keyword{package}
