% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HiCDOC}
\alias{HiCDOC}
\title{Default pipeline to run on a \code{\link{HiCDOCDataSet}}.}
\usage{
HiCDOC(object, parallel = TRUE)
}
\arguments{
\item{object}{A \code{\link{HiCDOCDataSet}}.}

\item{parallel}{Whether or not to parallelize each step. Defaults to TRUE.}
}
\value{
A \code{\link{HiCDOCDataSet}} with all slots filled.
}
\description{
Runs the default filtering, normalization, and computational steps on a
\code{\link{HiCDOCDataSet}}. To learn more about HiCDOC, browse the vignette:
\code{browseVignettes(package = "HiCDOC")}.
}
\details{
\subsection{\code{HiCDOCDataSet} object}{
    A \code{\link{HiCDOCDataSet}} object can be created from different files:
    \itemize{
        \item{Tabular files: }{
            see \code{\link{HiCDOCDataSetFromTabular}}
        }
        \item{(m)Cool files: }{
            see \code{\link{HiCDOCDataSetFromCool}}
        }
        \item{HiC files: }{
            see \code{\link{HiCDOCDataSetFromHiC}}
        }
        \item{HiC-Pro matrices and bed files: }{
            see \code{\link{HiCDOCDataSetFromHiCPro}}
        }
    }
}
\subsection{\code{HiCDOC} pipeline}{
    The HiCDOC pipeline has seven steps:
    \describe{
        \item{Three filtering steps:}{
            \itemize{
                \item{\code{\link{filterSmallChromosomes}}}{
                    to filter out small chromosomes
                }
                \item{\code{\link{filterWeakPositions}}}{
                    to filter out weak positions with very few interactions
                }
                \item{\code{\link{filterSparseReplicates}}}{
                    to filter out sparse replicates with many null
                    interactions
                }
            }
        }
        \item{Three normalization steps:}{
            \itemize{
                \item{\code{\link{normalizeTechnicalBiases}}}{
                    to normalize technical biases in each replicates
                }
                \item{\code{\link{normalizeBiologicalBiases}}}{
                    to normalize biological biases in each replicate
                }
                \item{\code{\link{normalizeDistanceEffect}}}{
                    to normalize the distance effect in each chromosome
                }
            }
        }
        \item{One computational step:}{
            \itemize{
                \item{\code{\link{detectCompartments}}}{
                    to detect compartments in each condition and find
                    significant changes between conditions.
                }
            }
        }
    }
}
}
\examples{
object <- HiCDOCDataSetExample()
object <- HiCDOC(object)

# Equivalent to
\dontrun{
object <- filterSmallChromosomes(object)
object <- filterSparseReplicates(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
}

}
\seealso{
\code{\link{HiCDOCDataSet}}
}
