% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/HiCDOC-package.R
\docType{package}
\name{HiCDOC}
\alias{HiCDOC}
\alias{HiCDOC-package}
\title{Main method.  Start the pipeline with default parameters.}
\usage{
HiCDOC(object)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}
}
\description{
\code{HiCODC} is a package...
TODO
}
\examples{

}
\author{
Cyril Kurylo and Matthias Zytnicki and Sylvain Foissac and Élise Maigné
}
\keyword{package}
