% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods.R
\docType{class}
\name{HiCDOCDataSet-class}
\alias{HiCDOCDataSet-class}
\alias{HiCDOCDataSet}
\alias{chromosomes}
\alias{conditions}
\alias{replicates}
\alias{interactions}
\alias{positions}
\alias{differences}
\alias{concordances}
\alias{compartments}
\alias{centroids}
\alias{show}
\alias{chromosomes,HiCDOCDataSet-method}
\alias{conditions,HiCDOCDataSet-method}
\alias{replicates,HiCDOCDataSet-method}
\alias{interactions,HiCDOCDataSet-method}
\alias{positions,HiCDOCDataSet-method}
\alias{differences,HiCDOCDataSet-method}
\alias{concordances,HiCDOCDataSet-method}
\alias{compartments,HiCDOCDataSet-method}
\alias{centroids,HiCDOCDataSet-method}
\alias{show,HiCDOCDataSet-method}
\alias{parameters,HiCDOCDataSet-method}
\title{Infrastructure for HiCDOC experiment and differential interaction.}
\usage{
\S4method{chromosomes}{HiCDOCDataSet}(object)

\S4method{conditions}{HiCDOCDataSet}(object)

\S4method{replicates}{HiCDOCDataSet}(object)

\S4method{interactions}{HiCDOCDataSet}(object)

\S4method{positions}{HiCDOCDataSet}(object)

\S4method{differences}{HiCDOCDataSet}(object)

\S4method{concordances}{HiCDOCDataSet}(object)

\S4method{compartments}{HiCDOCDataSet}(object)

\S4method{centroids}{HiCDOCDataSet}(object)

\S4method{show}{HiCDOCDataSet}(object)

\S4method{parameters}{HiCDOCDataSet}(object)
}
\arguments{
\item{object}{a HiCDOCDataSet object.}
}
\value{
For \code{chromosomes}: a character vector, with the name 
of chromosomes

For \code{conditions}: a character vector, with the name of 
conditions, with repetition over the replicates.

For \code{replicates}: a character vector, with the name of 
conditions, with repetition over the conditions

For \code{interactions}: a tibble of the interactions on all 
the choromosomes, conditions and replicates

For \code{positions}: a tibble of the positions corresponding
to the bins, for each chromosome

For \code{differences}: a tibble of the differences found by
\code{detectCompartments}

For \code{concordances}: a tibble of the concordances found by
\code{detectCompartments}

For \code{compartments}: a tibble of the compartments found by
\code{detectCompartments}

For \code{centroids}: a tibble of the centroids found by
\code{detectCompartments}
}
\description{
\code{HiCDOCDataSet} is an S4 class providing the infrastructure (slots)
to store the input data, methods parameters, intermediate calculations
and results of a differential interaction pipeline
}
\details{
A HiCDOCDataSet can be constructed from 4 different types of data :
\itemize{
\item Matrices: see \code{\link{HiCDOCDataSetFromSparseMatrix}}
\item Cool or mCool data: see \code{\link{HiCDOCDataSetFromCool}}
\item Hic files: see \code{\link{HiCDOCDataSetFromHic}}
\item Hic Pro matrices and bed files: see \code{\link{HiCDOCDataSetFromHicPro}}
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{chromosomes}: access and print the chromosomes

\item \code{conditions}: access and print the conditions

\item \code{replicates}: access and print the replicates

\item \code{interactions}: access and print the interactions matrix

\item \code{positions}: access and print the positions

\item \code{differences}: access and print the differences

\item \code{concordances}: access and print the concordances

\item \code{compartments}: access and print the compartments

\item \code{centroids}: access and print the centroids

\item \code{show}: access and print the object

\item \code{parameters}: access and print the parameters
}}

\section{Slots}{

\describe{
\item{\code{inputPath}}{The names of the matrix input files.}

\item{\code{interactions}}{The interaction matrices.}

\item{\code{weakBins}}{The empty bins.}

\item{\code{chromosomes}}{The list of chromosomes.}

\item{\code{replicates}}{The names of the replicates.}

\item{\code{totalReplicates}}{The names of the replicates, glued with the
name of the conditions.}

\item{\code{totalReplicatesPerCondition}}{A 2-element list, one for each condition,
where the union is the names of the replicates.}

\item{\code{conditions}}{The names of the conditions (exactly two different).}

\item{\code{totalBins}}{The number of bins per chromosome.}

\item{\code{binSize}}{The resolution.}

\item{\code{distances}}{The distribution of distances to the centroids.}

\item{\code{diagonalRatios}}{TODO}

\item{\code{compartments}}{The A/B compartments distribution, along the chromosomes.}

\item{\code{concordances}}{The concordance distribution, along the chromosomes.}

\item{\code{differences}}{The distribution of the difference of the concordance.}

\item{\code{centroids}}{The position of the centroids.}

\item{\code{parameters}}{An named \code{list}. The parameters for the
segmentation methods. See \code{\link{parameters}}.}

\item{\code{positions}}{The position of the bins.}
}}

\examples{
object <- HiCDOCExample()
## Print the object
object

## Accessors
chromosomes(object)
conditions(object)
replicates(object)
interactions(object)
positions(object)
parameters(object)

## Accessors after a run of detectCompartments
object <- runHiCDOC(object)
differences(object)
concordances(object)
compartments(object)
centroids(object)
}
\seealso{
\code{\link{HiCDOCExample}},
\code{\link{parameters}},
\code{\link{HiCDOCDataSetFromSparseMatrix}},
\code{\link{HiCDOCDataSetFromCool}},
\code{\link{HiCDOCDataSetFromHic}},
\code{\link{HiCDOCDataSetFromHicPro}}

\code{\link[HiCDOC]{parameters}} function to access the parameters of a 
HiCDOCDataSet object.
}
