% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSet-class}
\alias{HiCDOCDataSet-class}
\alias{HiCDOCDataSet}
\title{\code{HiCDOCDataSet} S4 class.}
\description{
Data structure for a Hi-C experiment.
}
\details{
An instance of \code{HiCDOCDataSet} describes a Hi-C experiment with slots
for path(s) to input file(s), interactions, pipeline parameters defaulting to
\code{defaultHiCDOCParameters}, and computation results. It can be
constructed from 4 different types of data:
\itemize{
\item Tabular files: see \code{\link{HiCDOCDataSetFromTabular}}
\item (m)Cool files: see \code{\link{HiCDOCDataSetFromCool}}
\item HiC files: see \code{\link{HiCDOCDataSetFromHiC}}
\item HiC-Pro matrices and bed files: see \code{\link{HiCDOCDataSetFromHiCPro}}
An example \code{HiCDOCDataSet} is also available, see
\code{\link{exampleHiCDOCDataSet}}.
}
}
\section{Slots}{

\describe{
\item{\code{input}}{A vector of path(s) to input file(s).}

\item{\code{parameters}}{A list of parameters used for filtering, normalization, and prediction of
compartments.}

\item{\code{interactions}}{An InteractionSet object of interactions.}

\item{\code{chromosomes}}{A vector of names of chromosomes.}

\item{\code{totalBins}}{A list of the number of bins in each chromosome.}

\item{\code{weakBins}}{A list of weak bins that are filtered out in each chromosome.}

\item{\code{validAssay}}{A list of non-sparse valid conditions and replicates, corresponding to the
valid columns of the assay matrix, for each chromosome}

\item{\code{compartments}}{A tibble of the A or B compartment of each bin in each condition.}

\item{\code{concordances}}{A tibble of the concordance of each bin in each replicate.}

\item{\code{differences}}{A tibble of detected compartment differences between conditions.}

\item{\code{comparisons}}{A tibble of comparisons ??? TODO}

\item{\code{distances}}{A tibble of the distances to centroids of each bin in each replicate.}

\item{\code{centroids}}{A tibble of centroids in each chromosome and condition.}

\item{\code{selfInteractionRatios}}{A tibble of differences between self interaction and other interactions for
each bin in each replicate.}
}}

\seealso{
\code{\link{HiCDOC}}
\code{\link{exampleHiCDOCDataSet}},
\code{\link{HiCDOCDataSetFromTabular}},
\code{\link{HiCDOCDataSetFromCool}},
\code{\link{HiCDOCDataSetFromHiC}},
\code{\link{HiCDOCDataSetFromHiCPro}}
}
