% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{HiCDOCDataSet-methods}
\alias{HiCDOCDataSet-methods}
\alias{chromosomes}
\alias{positions}
\alias{conditions}
\alias{replicates}
\alias{resolution}
\alias{interactions}
\alias{compartments}
\alias{concordances}
\alias{differences}
\alias{show}
\alias{show,HiCDOCDataSet-method}
\title{Methods to access a \code{\link{HiCDOCDataSet}} components.}
\usage{
chromosomes(object)

positions(object)

conditions(object)

replicates(object)

resolution(object)

compartments(object)

differences(object, threshold = 0.05)

concordances(object)

object
}
\arguments{
\item{object}{a HiCDOCDataSet object}

\item{threshold}{a numeric value between 0 and 1. If no threshold, all the differences will
be printed even the non significant ones. Otherwise the differences printed
are filtered to show the ones with an adjusted p-value <= \code{threshold}.}
}
\value{
A character vector (for \code{chromosomes}, \code{conditions},
\code{replicates}), an integer(for \code{resolution}), a tibble
(for \code{interactions} and \code{positions}), or a GRanges object
(for \code{compartments}, \code{concordances}, \code{differences}).
}
\description{
Retrieve information and results from a \code{\link{HiCDOCDataSet}}.
}
\section{Functions}{
\itemize{
\item \code{chromosomes}: Retrieves the vector of chromosome names.

\item \code{positions}: Retrieves the genomic positions corresponding to bins for each chromosome.

\item \code{conditions}: Retrieves the vector of condition names.

\item \code{replicates}: Retrieves the vector of replicate names.

\item \code{interactions}: Retrieves a tibble of the interactions.

\item \code{resolution}: Retrieves the resolution (span of each position in number of bases).

\item \code{compartments}: Retrieves a \code{GenomicRange} of the compartment of every position
in every condition.

\item \code{differences}: Retrieves a \code{GenomicRange} of the significant compartment differences
between conditions, and their p-values.

\item \code{concordances}: Retrieves a \code{GenomicRange} of the concordance (confidence in assigned
compartment) of every position in every replicate.

\item \code{show,HiCDOCDataSet-method}: Describes the object and its methods.
}}

\examples{
object <- HiCDOCDataSetExample()
chromosomes(object)
conditions(object)
object <- HiCDOC(object)
differences(object)
}
