% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSet}
\alias{HiCDOCDataSet}
\alias{HiCDOCDatSet}
\alias{HiCDOCDataSet-class}
\title{Infrastructure for HiCDOC data set}
\description{
\code{HiCDOCDataSet} is an S4 class providing the infrastructure (slots)
to store the input data.
}
\details{
\code{HiCDOCDataSet} does this and that...
TODO
}
\section{Slots}{

\describe{
\item{\code{inputPath}}{The input matrix/matrices.  Can be the path(s) to
to one, or several files, depending on the format considered.}

\item{\code{interactions}}{A data frame storing the matrices.  Is built using
the values in the \code{inputPath} files.}

\item{\code{replicates}}{A vector storing the names of the replicates.}

\item{\code{conditions}}{A vector storing the names of the conditions.
There should be exactly two different conditions.}

\item{\code{binSize}}{The resolution.}
}}

