% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSet}
\alias{HiCDOCDataSet}
\alias{HiCDOCDataSet-class}
\title{Infrastructure for HiCDOC experiment and differential interaction}
\usage{
HiCDOCDataSet(object = NULL, parameters = NULL, binSize = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{parameters}{A named \code{list}. The parameters for the
segmentation methods. See \code{\link{parameters}}.}

\item{binSize}{The resolution.}
}
\value{
\code{HiCDOCDataSet} constructor returns an \code{HiCDOCDataSet}
        object of class S4.
}
\description{
\code{HiCDOCDataSet} is an S4 class providing the infrastructure (slots)
to store the input data, methods parameters, intermediate calculations
and results of a differential interaction pipeline
}
\details{
HiCDOCDataSet does this
}
\section{Slots}{

\describe{
\item{\code{inputPath}}{The names of the matrix input files.}

\item{\code{interactions}}{The interaction matrices.}

\item{\code{weakBins}}{The empty bins.}

\item{\code{chromosomes}}{The list of chromosomes.}

\item{\code{replicates}}{The names of the replicates.}

\item{\code{totalReplicates}}{The names of the replicates, glued with the
name of the conditions.}

\item{\code{totalReplicatesPerCondition}}{A 2-element list, one for each condition,
where the union is the names of the replicates.}

\item{\code{conditions}}{The names of the conditions (exactly two different).}

\item{\code{totalBins}}{The number of bins per chromosome.}

\item{\code{binSize}}{The resolution.}

\item{\code{distances}}{The distribution of distances to the centroids.}

\item{\code{diagonalRatios}}{TODO}

\item{\code{compartments}}{The A/B compartments distribution, along the chromosomes.}

\item{\code{concordances}}{The concordance distribution, along the chromosomes.}

\item{\code{differences}}{The distribution of the difference of the concordance.}

\item{\code{centroids}}{The position of the centroids.}

\item{\code{parameters}}{An named \code{list}. The parameters for the
segmentation methods. See \code{\link{parameters}}.}
}}

\examples{
linkToMatrix <- system.file("extdata", "sampleMatrix.tsv", package="HiCDOC")
srnaDataSet <- HiCDOCDataSetFromSparseMatrix(linkToMatrix)
srnaExp <- HiCDOCDataSet(srnaDataSet)
}
