% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HiCDOCDataSetExample}
\alias{HiCDOCDataSetExample}
\title{Example of HiCDOCDataSet data.}
\format{
S4 HiCDOCDataSet object with the following characteristics :
\describe{
  \item{chromosomes}{4 chromosomes : "W", "X", "Y" and "Z"}
  \item{replicates}{3 differents replicates : R1, R2 and R3}
  \item{conditions}{3 differents conditions : 1, 2, 3}
  \item{resolution}{the computed resolution is 137.}
}
}
\usage{
HiCDOCDataSetExample
}
\value{
A HiCDOCDataSet object
}
\description{
A S4 HiCDOCDataSet object with 4 chromosomes, 2 conditions and 3 replicates.
}
\keyword{datasets}
