% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSetFromCool}
\alias{HiCDOCDataSetFromCool}
\title{Construct a \code{HiCDOCDataSet} from a cool file.}
\usage{
HiCDOCDataSetFromCool(coolFileNames, replicates, conditions)
}
\arguments{
\item{coolFileNames}{A vector of file names, each one being a .cool file.}

\item{replicates}{character vector. The list of the replicate names.
This vector can have repeated values if the replicates are repeated 
over the conditions.}

\item{conditions}{character vector. The names of the two conditions. 
This should be repeated values of the names, the vector should have the 
same size as \code{replicates}.}
}
\value{
\code{HiCDOCDataSet} constructor returns an \code{HiCDOCDataSet}
        object of class S4.
}
\description{
Construct a \code{HiCDOCDataSet} from a cool file.
}
\examples{
basedir <- system.file("extdata", package="HiCDOC")
data    <- read.csv(file.path(basedir, "coolData.csv"))
data
\dontrun{
dataSet <- HiCDOCDataSetFromCool(file.path(basedir, data$FileName),
                                 data$Replicate,
                                 data$Condition)
}

}
