% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSetFromHic}
\alias{HiCDOCDataSetFromHic}
\title{Construct a \code{HiCDOCDataSet} from a hic file.}
\usage{
HiCDOCDataSetFromHic(hicFileNames, replicates, conditions, resolution)
}
\arguments{
\item{hicFileNames}{A vector of file names, each one being a .hic file.}

\item{replicates}{The list of the replicate names.}

\item{conditions}{The names of the two conditions.}

\item{resolution}{The resolution.}
}
\value{
\code{HiCDOCDataSet} constructor returns an \code{HiCDOCDataSet}
        object of class S4.
}
\description{
Construct a \code{HiCDOCDataSet} from a hic file.
}
