% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HiCDOCDataSetFromHic}
\alias{HiCDOCDataSetFromHic}
\title{Construct a \code{HiCDOCDataSet} from a hic file.}
\usage{
HiCDOCDataSetFromHic(hicFileNames, replicates, conditions, resolution)
}
\arguments{
\item{hicFileNames}{A vector of file names, each one being a .hic file.}

\item{replicates}{character vector. The list of the replicate names.
This vector can have repeated values if the replicates are repeated 
over the conditions.}

\item{conditions}{character vector. The names of the two conditions. 
This should be repeated values of the names, the vector should have the 
same size as \code{replicates}.}

\item{resolution}{integer. The resolution of the matrices.}
}
\value{
\code{HiCDOCDataSet} constructor returns an \code{HiCDOCDataSet}
        object of class S4.
}
\description{
Construct a \code{HiCDOCDataSet} from a hic file.
}
