% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HiCDOCDataSetFromHiCPro}
\alias{HiCDOCDataSetFromHiCPro}
\title{\code{\link{HiCDOCDataSet}} constructor from HiC-Pro files.}
\usage{
HiCDOCDataSetFromHiCPro(matrixPaths, bedPaths, replicates, conditions)
}
\arguments{
\item{matrixPaths}{A vector of paths to HiC-Pro matrix files.}

\item{bedPaths}{A vector of paths to HiC-Pro bed files.}

\item{replicates}{A vector of replicate names repeated along the conditions.}

\item{conditions}{A vector of condition names repeated along the replicates.}
}
\value{
A \code{\link{HiCDOCDataSet}}.
}
\description{
Constructs a \code{\link{HiCDOCDataSet}} from a set of HiC-Pro generated
files.
}
\examples{
# Retrieve HiC-Pro files
directory <- system.file("extdata", package = "HiCDOC")
matrixPaths <- list.files(directory, '\\\\.matrix$')[0:5]
bedPaths <- list.files(directory, '\\\\.bed$')[0:5]
# Specify replicate and condition for each pair of files
# In this case:
# - The first files in the paths vectors are replicate 1 of condition a
# - The second files in the paths vectors are replicate 2 of condition a
# - ...
# - The last files in the paths vectors are replicate x of condition 3
replicates <- c(  1,   2,   1,   2, 'x')
conditions <- c('a', 'a',   2,   2,   3)
# Create HiCDOC object
object <- HiCDOCDataSetFromHiCPro(
    matrixPaths,
    bedPaths,
    replicates = replicates,
    conditions = conditions
)

}
