% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HiCDOCExample}
\alias{HiCDOCExample}
\title{Example constructor}
\usage{
HiCDOCExample()
}
\value{
An \code{HiCDOCDataSet} object called '\code{exp}'.
}
\description{
This function provides an example of a \code{HiCDOCDataSet} object from
Marti-Marimon et al.
}
\examples{
## The 'HiCDOCDataSet' object in this example was constructed by:
exp <- HiCDOCExample()
exp

}
