% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HiCDOCExample}
\alias{HiCDOCExample}
\title{Example \code{\link{HiCDOCDataSet}} constructor.}
\usage{
HiCDOCExample()
}
\value{
A \code{\link{HiCDOCDataSet}}.
}
\description{
Constructs a toy \code{\link{HiCDOCDataSet}} from simulated data.
}
\examples{
object <- HiCDOCExample()

}
