% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{interactions}
\alias{interactions}
\alias{interactions,HiCDOCExp-method}
\title{Accessors for the 'interactions' slot of an HiCDOCExp object}
\usage{
\S4method{interactions}{HiCDOCExp}(object)
}
\arguments{
\item{object}{An \code{HiCDOCExp} object.}
}
\value{
A tibble
}
\description{
The \code{interactions} slot contains the (transformed) interaction
profiles.
}
\examples{
exp <- HiCDOCExample()
interactions(exp)

}
