% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeDistanceEffect.R
\name{normalizeDistanceEffect}
\alias{normalizeDistanceEffect}
\title{Normalize the distance effect using a loess on the intensity vs distance
to diagonal.}
\usage{
normalizeDistanceEffect(object, sampleSize = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{sampleSize}{A numerical value. The number of bins used when sampling
all the bins. If NULL, default to the first not NULL of 
\code{object$sampleSize} and \code{HiCDOCDefaultParameters$sampleSize}.}
}
\value{
A \code{HiCDOCDataSet} object, with the normalized matrices.
}
\description{
Normalize the distance effect using a loess on the intensity vs distance
to diagonal.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
}
