% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeDistanceEffect.R
\name{normalizeDistanceEffect}
\alias{normalizeDistanceEffect}
\title{Normalize the distance effect on a HiCDOCDataSet object using a loess
on the intensity vs distance to diagonal.}
\usage{
normalizeDistanceEffect(object, loessSampleSize = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{loessSampleSize}{A numerical value. The number of bins used when sampling
all the bins. If NULL, default to the first not NULL of
\code{object$loessSampleSize} and \code{defaultHiCDOCParameters$loessSampleSize}.}
}
\value{
A \code{HiCDOCDataSet} object, with the normalized matrices.
}
\description{
Normalize the distance effect on a HiCDOCDataSet object using a loess
on the intensity vs distance to diagonal.
}
\examples{
object <- HiCDOCExample()
object <- normalizeDistanceEffect(object)
}
\seealso{
\code{\link[HiCDOC]{normalizeTechnicalBiases}},
\code{\link[HiCDOC]{normalizeBiologicalBiases}} and
\code{\link[HiCDOC]{HiCDOC}} for the recommended pipeline.
}
