% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeTechnicalBiases.R
\name{normalizeTechnicalBiases}
\alias{normalizeTechnicalBiases}
\title{Normalize technical biases.}
\usage{
normalizeTechnicalBiases(object, parallel = TRUE)
}
\arguments{
\item{object}{A \code{\link{HiCDOCDataSet}}.}

\item{parallel}{Whether or not to parallelize the processing. Defaults to TRUE.}
}
\value{
A \code{\link{HiCDOCDataSet}} with normalized interactions.
}
\description{
Normalizes technical biases such as sequencing depth by using a cyclic loess
to recursively normalize each pair of interaction matrices. Depends on
\code{multiHiCcompare}.
}
\examples{
object <- HiCDOCDataSetExample()
object <- filterSparseReplicates(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)

}
\seealso{
\code{\link{filterSparseReplicates}},
\code{\link{filterWeakPositions}},
\code{\link{normalizeBiologicalBiases}},
\code{\link{normalizeDistanceEffect}},
\code{\link{HiCDOC}}
}
