% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeTechnicalBiases.R
\name{normalizeTechnicalBiases}
\alias{normalizeTechnicalBiases}
\title{Normalize the distance effect using a cyclic loess on all the matrices.}
\usage{
normalizeTechnicalBiases(object, parallel = FALSE)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{parallel}{Logical, defautl to FALSE. Should parallel computing be 
used in \code{\link[multiHiCcompare]{cyclic_loess}} ?}
}
\value{
A \code{HiCDOCDataSet} object, with the normalized matrices.
}
\description{
Normalize the distance effect using a cyclic loess on all the matrices.
}
\examples{
object <- HiCDOCExample()
object <- normalizeTechnicalBiases(object)
}
\seealso{
\code{\link[HiCDOC]{normalizeBiologicalBiases}}, 
\code{\link[HiCDOC]{normalizeDistanceEffect}} and 
\code{\link[HiCDOC]{HiCDOC}} for the recommended pipeline.
}
