% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,HiCDOCDataSet-method}
\alias{parameters<-}
\alias{parameters<-,HiCDOCDataSet-method}
\title{Accessors for the 'parameters' slot of an HiCDOCDataSet object}
\usage{
\S4method{parameters}{HiCDOCDataSet}(object)

\S4method{parameters}{HiCDOCDataSet}(object) <- value
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}

\item{value}{A named \code{list} containing valid parameters. See details.}
}
\value{
The named list of the parameters used in the analysis.
}
\description{
The \code{parameters} slot holds the parameter values
used in an experiment as a named \code{list}. Default values
exist for parameters, but these can also be supplied as input
values in the \code{useParameters} argument of the \code{\link{HiCDOC}}
function or using the assignment function \code{\link{parameters<-}}.
}
\details{
Parameters in a HiCDOC experiment.

\subsection{Global parameters}{
   \describe{
       \item{\code{minLengthChr}}{The minimum chromosome
               size (in number of bins), to be kept by the function
               \code{filterSmallChromosomes()}. Default to 100.}
       \item{\code{weakPosThreshold}}{To be kept by the function
               \code{filterWeakPositions()}, the bins (positions)
               of a chromosome must have a mean value greater than
               \code{weakPosThreshold}, on all replicates and all
               conditions. The mean is computed on the row of the
               reconstructed full interaction matrix for 1 chromosome,
               1 condition and 1 replicate. Default to 0.}
       \item{\code{sparseThreshold}}{To be kept by the function
               \code{filterSparseChromosomes()}, the sparsity (percentage
               of empty cells) of the interactions matrix must be lower
               than the threshold, on all replicates and all conditions.
               Default to 0.95.}
      \item{\code{sampleSize}}{The number of bins used when sampling
               all the bins in \code{normalizeDistanceEffect()}.
               Default to 20000.}
      \item{\code{kMeansIterations}}{The maximum number of 2-means
               iterations.}
      \item{\code{kMeansDelta}}{The stop criterion of convergence of
               the 2-means method.}
      \item{\code{kMeansRestarts}}{The maximum number of restarts
               for the 2-means.}
   }
}
}
\examples{
exp <- HiCDOCExample()
parameters(exp)
parameters(exp) <- list("weakPosThreshold" = 100)
}
\seealso{
\code{useParameters} argument in \code{\link{HiCDOC}} function.
}
