% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,HiCDOCDataSet-method}
\alias{parameters<-}
\alias{parameters<-,HiCDOCDataSet-method}
\alias{printHiCDOCParameters}
\title{Accessors for the 'parameters' slot of an HiCDOCDataSet object}
\usage{
\S4method{parameters}{HiCDOCDataSet}(object)

\S4method{parameters}{HiCDOCDataSet}(object) <- value

printHiCDOCParameters(object)
}
\arguments{
\item{object}{a HiCDOCDataSet object}

\item{value}{A named \code{list} containing valid parameters. See details.}
}
\value{
The named list of the parameters used in the analysis.
}
\description{
The \code{parameters} slot holds the parameter values
used in an experiment as a named \code{list}. Default values
exist for parameters, but these can also be supplied as input
values in the \code{useParameters} argument of the \code{\link{HiCDOC}}
function or using the assignment function \code{\link{parameters<-}}.
}
\details{
Parameters in a HiCDOC experiment.

\subsection{Global parameters}{
   \describe{
       \item{\code{minDepth}}{The cutoff to filter the base-level
               coverage. Bases where at least one sample has (normalized)
               coverage greater than \code{minDepth} be been retained.
               Default to \code{10}.}
      \item{\code{sampleSize}}{The number of bins used when sampling
               all the bins.}
      \item{\code{loessSpan}}{The optimal span value used for the
               diagonal normalization.}
      \item{\code{kMeansIterations}}{The maximum number of 2-means
               iterations.}
      \item{\code{kMeansDelta}}{The stop criterion of convergence of
               the 2-means method.}
      \item{\code{kMeansRestarts}}{The maximum number of restarts
               for the 2-means.}
      \item{\code{minLengthChr}}{The minimum chromosome
               size (in number of bins), to be kept.}
      \item{\code{weakPosThreshold}}{To be kept, the bins (positions)
               of a chromosome must have a mean value greater than
               \code{weakPosThreshold}, on all replicates and all
               conditions. The mean is computed on the row of the
               reconstructed full interaction matrix for 1 chromosome,
               1 condition and 1 replicate.}
   }
}
}
\examples{
exp <- HiCDOCExample()
exp <- HiCDOC(exp)
print(parameters(exp))

exp <- HiCDOCExample()
exp <- HiCDOC(exp)
print(parameters(exp))

}
\seealso{
\code{useParameters} argument in \code{\link{HiCDOC}} function.
}
