% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseInteractionMatrix3Columns}
\alias{parseInteractionMatrix3Columns}
\title{Read interaction matrices in tabular 3-column format.}
\usage{
parseInteractionMatrix3Columns(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
object An \code{HiCDOCDataSet} object.
}
\description{
This function parses the tsv files provided in
\code{object@inputMatrixPath}.
}
\examples{
\dontrun{
linkToMatrix <-system.file("extdata", "sampleMatrix.tsv", package = "HiCDOC")
data   <- makeHiCDOCDataSet(inputPath = linkToMatrix)
object <- parseInteractionMatrix3Columns(data)
}
}
