% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseInteractionMatrixCool}
\alias{parseInteractionMatrixCool}
\title{Read interaction matrices in .cool format.}
\usage{
parseInteractionMatrixCool(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
object An \code{HiCDOCDataSet} object.
}
\description{
This function parses the .cool files provided in
\code{object@inputMatrixPath}.
}
