% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseInteractionMatrixHic}
\alias{parseInteractionMatrixHic}
\title{Read interaction matrices in .hic format.}
\usage{
parseInteractionMatrixHic(object)
}
\arguments{
\item{object}{An \code{HiCDOCDataSet} object.}
}
\value{
object An \code{HiCDOCDataSet} object.
}
\description{
This function parses the .hic files provided in
\code{object@inputMatrixPath}.
}
