% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotAB}
\alias{plotAB}
\title{Plot the distribution of A/B compartments along the genomic positions.}
\usage{
plotAB(object, chromosomeId)
}
\arguments{
\item{object}{an \code{HiCDOCDataSet} object}

\item{chromosomeId}{character or numeric value, name or number of chromosome}

\item{conditionId}{character or numeric value, name or number of condition}
}
\value{
A list of \code{ggplot}, one for each chromosome.
}
\description{
Plot the distribution of A/B compartments along the genomic positions.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotAB(object, 1, 1)
}
