% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCentroids.R
\name{plotCentroids}
\alias{plotCentroids}
\title{Plot the centroids PCA for a given chromosome.}
\usage{
plotCentroids(object, chromosome, size = 2)
}
\arguments{
\item{object}{an \code{HiCDOCDataSet} object}

\item{size}{Numeric. Size of the points in geom_point.}

\item{chromosomeId}{Character or numeric value. Name or number of
the chromosome, like in chromosomes(object)}
}
\value{
A \code{ggplot} object
}
\description{
Plot the centroids PCA for a given chromosome.
}
\examples{
object <- HiCDOCDataSetExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotCentroids(object, 1)
}
