% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCentroids}
\alias{plotCentroids}
\title{Plot the centroid distributions along the genomic positions
for a given chromosome.}
\usage{
plotCentroids(object, chromosomeId)
}
\arguments{
\item{object}{an \code{HiCDOCDataSet} object}

\item{chromosomeId}{Character or numeric value. Name or number of
the chromosome, like in chromosomes(object)}
}
\value{
A \code{ggplot} object
}
\description{
Plot the centroid distributions along the genomic positions
for a given chromosome.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotCentroids(object, 1)
}
