% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCentroids}
\alias{plotCentroids}
\title{Plot the centroid distributions along the genomic positions.}
\usage{
plotCentroids(object)
}
\arguments{
\item{objet}{an \code{HiCDOCExp} object}
}
\value{
A list of \code{ggplot}, one for each chromosome.
}
\description{
Plot the centroid distributions along the genomic positions.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotCentroids(object)
}
