% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCentroids.R
\name{plotCentroids}
\alias{plotCentroids}
\title{Plot centroids.}
\usage{
plotCentroids(object, chromosome, size = 2)
}
\arguments{
\item{object}{A \code{\link{HiCDOCDataSet}}.}

\item{chromosome}{A chromosome name or index in \code{chromosomes(object)}.}

\item{size}{Size of each point. Defaults to 2.}
}
\value{
A \code{ggplot}.
}
\description{
Plots the result of the PCA on the compartments' centroids.
}
\examples{
data(HiCDOCDataSetExample)
object <- HiCDOC(HiCDOCDataSetExample)
plotCentroids(object, chromosome = 1)

}
