% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompartmentChanges.R
\name{plotConcordance}
\alias{plotConcordance}
\title{Plot the concordance}
\usage{
plotConcordance(
  object,
  chromosomeId,
  xlim = NULL,
  padjThreshold = 0.05,
  points = FALSE
)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object on which \code{detectCompartments()} has run.}

\item{chromosomeId}{The name or number of the chromosome to plot.
If number, will be taken in \code{object@chromosomes[chromosomeId]}}

\item{xlim}{A numeric-value pair, indicating the interval of positions to represent.
Default to NULL = all positions.}

\item{padjThreshold}{Significance threshold for the changes. Default to 0.05.}

\item{points}{Logical (default to FALSE). If TRUE, points will be added on the concordance lines.}
}
\value{
A ggplot.
}
\description{
Plot the concordance of all the replicates for one chromosome.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotConcordance(object, 1)
}
