% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConcordances.R
\name{plotConcordances}
\alias{plotConcordances}
\title{Plot the concordance}
\usage{
plotConcordances(
  object,
  chromosome,
  xlim = NULL,
  threshold = 0.05,
  points = FALSE
)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object on which
\code{detectCompartments()} has run.}

\item{xlim}{A numeric-value pair, indicating the interval of positions
to represent.
Default to NULL = all positions.}

\item{threshold}{Significance threshold for the changes. Default to 0.05.}

\item{points}{Logical (default to FALSE). If TRUE, points will be added
on the concordance lines.}

\item{chromosomeId}{The name or number of the chromosome to plot.
If number, will be taken in \code{object@chromosomes[chromosomeId]}}
}
\value{
A ggplot.
}
\description{
Plot the concordance of all the replicates for one chromosome.
}
\examples{
object <- HiCDOCDataSetExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- detectCompartments(object)
plotConcordances(object, 1)
}
