% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDistanceEffect}
\alias{plotDistanceEffect}
\title{Plot the distance vs intensity matrix.}
\usage{
plotDistanceEffect(object)
}
\arguments{
\item{object}{an \code{HiCDOCDataSet} object}
}
\value{
A \code{ggplot}.
}
\description{
Plot the distance vs intensity matrix.
}
\examples{
object <- HiCDOCExample()
plotDistanceEffect(object)
}
