% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistanceEffect.R
\name{plotDistanceEffect}
\alias{plotDistanceEffect}
\title{Plot the distance effect.}
\usage{
plotDistanceEffect(object)
}
\arguments{
\item{object}{A \code{\link{HiCDOCDataSet}}.}
}
\value{
A \code{ggplot}.
}
\description{
Plots the distance effect on proportion of interactions.
}
\examples{
object <- HiCDOCDataSetExample()
plotDistanceEffect(object)

}
