% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractions.R
\name{plotInteractions}
\alias{plotInteractions}
\title{Plot the interaction matrix (as heatmap).}
\usage{
plotInteractions(
  object,
  chromosome,
  transform = "log2",
  colours = c("#000066", "#ffffbf", "#990000")
)
}
\arguments{
\item{object}{an \code{HiCDOCDataSet} object}

\item{chromosome}{The name or number of the chromosome to plot.
If number, will be taken in \code{object@chromosomes[chromosome]}}

\item{transform}{character: transformation of the color scale. Default to "log2".
See \code{\link[ggplot2]{scale_fill_gradientn}} for other accepted values.
Set to NULL for no transformation.}

\item{colours}{character vector, vector of colours to use for n-colour
gradient.
Default to \code{c("#000066", "#ffffbf", "#990000")}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the interaction matrix (as heatmap).
}
\examples{
object <- HiCDOCDataSetExample()
p <- plotInteractions(object, chromosome = 1, transform = "log2")
}
