% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractions.R
\name{plotInteractions}
\alias{plotInteractions}
\title{Plot interaction matrices.}
\usage{
plotInteractions(
  object,
  chromosome,
  transform = "log2",
  colours = c("#000066", "#ffffbf", "#990000")
)
}
\arguments{
\item{object}{A \code{\link{HiCDOCDataSet}}.}

\item{chromosome}{A chromosome name or index in \code{chromosomes(object)}.}

\item{transform}{Transformation of the color scale. Set to NULL for no transformation. See
\code{\link[ggplot2]{scale_fill_gradientn}} for other accepted values.
Defaults to "log2".}

\item{colours}{A character vector of colours to use for the gradient. See
\code{\link[ggplot2]{scale_fill_gradientn}} for more info. Defaults to
\code{c("#000066", "#ffffbf", "#990000")}.}
}
\value{
A \code{ggplot}.
}
\description{
Plots the interaction matrices as heatmaps.
}
\examples{
object <- HiCDOCDataSetExample()
p <- plotInteractions(object, chromosome = 1)

}
