% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractions.R
\name{plotInteractions}
\alias{plotInteractions}
\title{Plot interaction matrices.}
\usage{
plotInteractions(
  object,
  chromosome,
  transform = NULL,
  colours = c("#F6FFB8", "#FF00CC", "#310038")
)
}
\arguments{
\item{object}{A \code{\link{HiCDOCDataSet}}.}

\item{chromosome}{A chromosome name or index in \code{chromosomes(object)}.}

\item{transform}{Transformation of the color scale. Default to NULL (no transformation). See
\code{\link[ggplot2]{scale_fill_gradientn}} for other accepted values.}

\item{colours}{A character vector of colours to use for the gradient. See
\code{\link[ggplot2]{scale_fill_gradientn}} for more info. Defaults to
\code{c("#F6FFB8", "#FF00CC", "#310038")}.}
}
\value{
A \code{ggplot}.
}
\description{
Plots the interaction matrices as heatmaps.
}
\examples{
data(exampleHiCDOCDataSet)
plotInteractions(exampleHiCDOCDataSet, chromosome = 1)

}
