% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCompartments.R
\name{predictAB}
\alias{predictAB}
\title{Use ratio between diagonal and off-diagonal interactions to determine which
clusters correspond to compartments A and B.}
\usage{
predictAB(object, parallel = FALSE, nbCores = NULL)
}
\arguments{
\item{object}{A \code{HiCDOCDataSet} object.}

\item{parallel}{Logical. Should parallel processing be used?}

\item{nbCores}{Integer, the number of cores to use for parallelization.}
}
\value{
A \code{HiCDOCDataSet} object, with diagonalRatios, and with A and B
labels replacing cluster numbers in centroids, compartments, distances and
concordances.
}
\description{
Use ratio between diagonal and off-diagonal interactions to determine which
clusters correspond to compartments A and B.
}
\examples{
object <- HiCDOCExample()
object <- filterSmallChromosomes(object)
object <- filterWeakPositions(object)
object <- normalizeTechnicalBiases(object)
object <- normalizeBiologicalBiases(object)
object <- normalizeDistanceEffect(object)
object <- clusterize(object)
object <- predictAB(object)
}
