% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{reduceHiCDOCDataSet}
\alias{reduceHiCDOCDataSet}
\title{reduce an HiCDOCDataSet}
\usage{
reduceHiCDOCDataSet(
  object,
  chromosomes = NULL,
  conditions = NULL,
  replicates = NULL
)
}
\arguments{
\item{object}{and HiCDOCDataSet object}

\item{chromosomes}{default to NULL, chromosomes ID to keep}

\item{conditions}{default to NULL, conditions ID to keep}

\item{replicates}{default to NULL, replicates ID to keep}
}
\value{
a HiCDOCDataSET object, reduced by keeping only the chromosomes, 
conditions or replicates given in parameters
}
\description{
reduce an HiCDOCDataSet
}
\examples{
object <- HiCDOCExample()
objectReduced <-reduceHiCDOCDataSet(object, chromosome = "17", replicate = "1")
}
