% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{reduceHiCDOCDataSet}
\alias{reduceHiCDOCDataSet}
\title{reduce an HiCDOCDataSet}
\usage{
reduceHiCDOCDataSet(
  object,
  chromosomes = NULL,
  conditions = NULL,
  replicates = NULL,
  dropLevels = TRUE
)
}
\arguments{
\item{object}{and HiCDOCDataSet object}

\item{chromosomes}{default to NULL, chromosomes ID to keep}

\item{conditions}{default to NULL, conditions ID to keep}

\item{replicates}{default to NULL, replicates ID to keep}

\item{dropLevels}{Logical, default to TRUE. Chromosomes, conditions and 
replicates are in factor format, should the unused levels be removed ? 
It should be set to FALSE if reduced objects are meant to be re-combined 
later.}
}
\value{
a HiCDOCDataSET object, reduced by keeping only the chromosomes, 
conditions or replicates given in parameters
}
\description{
reduce an HiCDOCDataSet
}
\examples{
object <- HiCDOCExample()
objectSmall <-reduceHiCDOCDataSet(object, chromosome = "17", replicate = "1")
}
