% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validParameters.R
\name{testChromosome}
\alias{testChromosome}
\title{Test the existence of a given chromosome.}
\usage{
testChromosome(object, chromosomeId)
}
\arguments{
\item{object}{A \code{HiCDOCExp} object.}

\item{chromosomeId}{The condition name, or an error.}
}
\value{
The chromosome name or an error.
}
\description{
Test the existence of a given chromosome.
}
